--- src/react/KeybindingsScreenApp.tsx
@@ -1,375 +1,0 @@
-import { useState, useEffect, useRef } from 'react'
-import { AllKeyCodes } from 'contro-max/build/types/keyCodes'
-import { GamepadButtonName } from 'contro-max/build/gamepad'
-import { contro as controEx } from '../controls'
-import PixelartIcon from './PixelartIcon'
-import KeybindingsCustom from './KeybindingsCustom'
-import Button from './Button'
-import Screen from './Screen'
-import styles from './KeybindingsScreen.module.css'
-
-
-export default (
-  {
-    contro,
-    setBinding,
-    resetBinding,
-    isPS
-  }: {
-    contro: typeof controEx,
-    setBinding: (e, group, action, buttonNum, inputType?) => void,
-    resetBinding: (group, action, intputType) => void,
-    isPS?: boolean
-  }
-) => {
-  const containerRef = useRef<HTMLDivElement | null>(null)
-  const bindsMap = useRef({ keyboard: {} as any, gamepad: {} as any })
-  const [commands, setCommands] = useState(contro.inputSchema.commands)
-  const [userConfig, setUserConfig] = useState(contro.userConfig)
-  const [awaitingInputType, setAwaitingInputType] = useState(null as null | 'keyboard' | 'gamepad')
-  const [groupName, setGroupName] = useState('')
-  const [actionName, setActionName] = useState('')
-  const [buttonNum, setButtonNum] = useState(0)
-  const [, forceUpdate] = useState(false)
-
-  const handleClick = (group, action, index, type) => {
-    setAwaitingInputType(type)
-    setGroupName(prev => group)
-    setActionName(prev => action)
-    setButtonNum(prev => index)
-  }
-
-  const updateKeyboardBinding = (e) => {
-    if (!e.code || e.key === 'Escape' || !awaitingInputType) return
-    setBinding({ code: e.code, state: true }, groupName, actionName, buttonNum)
-    updateBindMap()
-    updateBindWarnings()
-  }
-
-  const updateGamepadBinding = (data) => {
-    if (!data.state && awaitingInputType) {
-      setAwaitingInputType(null)
-      return
-    }
-    if ('button' in data) {
-      contro.enabled = false
-      void Promise.resolve().then(() => { contro.enabled = true })
-      setBinding(data, groupName, actionName, buttonNum)
-    }
-
-    setAwaitingInputType(null)
-  }
-
-  const updateBindMap = () => {
-    bindsMap.current = { keyboard: {} as any, gamepad: {} as any }
-    if (commands) {
-      for (const [group, actions] of Object.entries(commands)) {
-        for (const [action, { keys, gamepadButtons }] of Object.entries(actions)) {
-          if (keys) {
-            let currKeys
-            if (userConfig?.[group]?.[action]?.keys) {
-              currKeys = userConfig[group][action].keys
-            } else {
-              currKeys = keys
-            }
-            for (const [index, key] of currKeys.entries()) {
-              bindsMap.current.keyboard[key] ??= []
-              if (!bindsMap.current.keyboard[key].some(obj => obj.group === group && obj.action === action && obj.index === index)) {
-                bindsMap.current.keyboard[key].push({ group, action, index })
-              }
-            }
-          }
-          if (gamepadButtons) {
-            let currButtons
-            if (userConfig?.[group]?.[action]?.gamepad) {
-              currButtons = userConfig[group][action].gamepad
-            } else {
-              currButtons = gamepadButtons
-            }
-            if (currButtons.length > 0) {
-              bindsMap.current.gamepad[currButtons[0]] ??= []
-              bindsMap.current.gamepad[currButtons[0]].push({ group, action, index: 0 })
-            }
-          }
-        }
-      }
-    }
-  }
-
-  const updateBindWarnings = () => {
-    if (!commands) return
-    for (const [group, actions] of Object.entries(commands)) {
-      for (const [action, { keys, gamepadButtons }] of Object.entries(actions)) {
-        if (!containerRef.current) continue
-        if (keys) {
-          let currKeys
-          if (userConfig?.[group]?.[action]?.keys) {
-            currKeys = userConfig[group][action].keys
-          } else {
-            currKeys = keys
-          }
-          for (const [index, key] of currKeys.entries()) {
-            const elem = containerRef.current.querySelector(`#bind-warning-${group}-${action}-keyboard-${index}`) as HTMLElement
-            if (!elem) continue
-            if (bindsMap.current.keyboard[key].length > 1) {
-              for (const bind of bindsMap.current.keyboard[key]) {
-                const currElem = containerRef.current.querySelector(`#bind-warning-${bind.group}-${bind.action}-keyboard-${bind.index}`) as HTMLElement
-                if (!currElem) continue
-                currElem.style.display = 'flex'
-              }
-            } else {
-              elem.style.display = 'none'
-            }
-          }
-          if (currKeys.length === 1) {
-            const elem = containerRef.current.querySelector(`#bind-warning-${group}-${action}-keyboard-1`) as HTMLElement
-            if (!elem) continue
-            elem.style.display = 'none'
-          }
-          if (currKeys.length === 0) {
-            for (const index of [0, 1]) {
-              const elem = containerRef.current.querySelector(`#bind-warning-${group}-${action}-keyboard-${index}`) as HTMLElement
-              if (!elem) continue
-              elem.style.display = 'none'
-            }
-          }
-        }
-        if (gamepadButtons) {
-          let currButtons
-          if (userConfig?.[group]?.[action]?.gamepad) {
-            currButtons = userConfig[group][action].gamepad
-          } else {
-            currButtons = gamepadButtons
-          }
-          const elem = containerRef.current.querySelector(`#bind-warning-${group}-${action}-gamepad-0`) as HTMLElement
-          if (!elem || !currButtons || currButtons.length === 0) continue
-          if (bindsMap.current.gamepad[currButtons[0]].length > 1) {
-            for (const bind of bindsMap.current.gamepad[currButtons[0]]) {
-              const currElem = containerRef.current.querySelector(`#bind-warning-${bind.group}-${bind.action}-gamepad-${bind.index}`) as HTMLElement
-              if (!currElem) continue
-              currElem.style.display = 'flex'
-            }
-          } else {
-            elem.style.display = 'none'
-          }
-        }
-      }
-    }
-  }
-
-  // fill binds map
-  useEffect(() => {
-    updateBindMap()
-  }, [])
-
-  useEffect(() => {
-    contro.on('pressedKeyOrButtonChanged', updateGamepadBinding)
-
-    return () => {
-      contro.off('pressedKeyOrButtonChanged', updateGamepadBinding)
-    }
-  }, [groupName, actionName])
-
-
-  return <Screen title="Keybindings" backdrop>
-    {awaitingInputType && <AwaitingInputOverlay isGamepad={awaitingInputType === 'gamepad'} />}
-    <div className={styles.container}
-      ref={containerRef}
-      onKeyDown={(e) => updateKeyboardBinding(e)}
-    >
-
-      {Object.entries(commands).map(([group, actions], index) => {
-        if (group === 'custom') return null
-        return <div key={`group-container-${group}-${index}`} className={styles.group}>
-          <div className={styles['group-category']}>{group}</div>
-          {Object.entries(actions).map(([action, { keys, gamepadButtons }]) => {
-            return <div key={`action-container-${action}`} className={styles.actionBinds}>
-              <div className={styles.actionName}>{parseActionName(action)}</div>
-              {
-                userConfig?.[group]?.[action]?.keys?.length ? <Button
-                  onClick={() => {
-                    setActionName(prev => action)
-                    setGroupName(prev => group)
-                    forceUpdate(prev => !prev)
-                    resetBinding(group, action, 'keyboard')
-                    updateBindMap()
-                    updateBindWarnings()
-                  }}
-                  className={styles['undo-keyboard']}
-                  icon={'pixelarticons:undo'}
-                />
-                  : null}
-
-
-
-              {[0, 1].map((key, index) => <ButtonWithMatchesAlert
-                key={`keyboard-${group}-${action}-${index}`}
-                group={group}
-                action={action}
-                index={index}
-                handleClick={handleClick}
-                inputType={'keyboard'}
-                parseBindingName={parseBindingName}
-                keys={keys}
-                userConfig={userConfig}
-                gamepadButtons={gamepadButtons}
-                isPS={isPS}
-              />)}
-              <ButtonWithMatchesAlert
-                key={`gamepad-${group}-${action}`}
-                group={group}
-                action={action}
-                index={0}
-                handleClick={handleClick}
-                parseBindingName={parseBindingName}
-                inputType={'gamepad'}
-                keys={keys}
-                userConfig={userConfig}
-                gamepadButtons={gamepadButtons}
-                isPS={isPS}
-              />
-              {
-                userConfig?.[group]?.[action]?.gamepad?.length ? <Button
-                  key={`keyboard-${group}-${action}`}
-                  onClick={() => {
-                    setActionName(prev => action)
-                    setGroupName(prev => group)
-                    forceUpdate(prev => !prev)
-                    resetBinding(group, action, 'gamepad')
-                    updateBindMap()
-                    updateBindWarnings()
-                  }}
-                  className={styles['undo-gamepad']}
-                  icon={'pixelarticons:undo'}
-                />
-                  : null
-              }
-            </div>
-          })}
-        </div>
-      })}
-
-      <KeybindingsCustom
-        commands={commands}
-        userConfig={userConfig}
-        awaitingInputType={awaitingInputType}
-        setBinding={setBinding}
-        setAwaitingInputType={setAwaitingInputType}
-        setGroupName={setGroupName}
-        setActionName={setActionName}
-        updateBindMap={updateBindMap}
-        updateBindWarnings={updateBindWarnings}
-        setButtonNum={setButtonNum}
-        handleClick={handleClick}
-        parseBindingName={parseBindingName}
-        resetBinding={resetBinding}
-        isPS={isPS}
-      />
-    </div>
-  </Screen>
-}
-
-export const ButtonWithMatchesAlert = ({
-  group,
-  action,
-  index,
-  handleClick,
-  inputType,
-  parseBindingName,
-  userConfig,
-  keys,
-  gamepadButtons,
-  isPS,
-}) => {
-  return <div
-    key={`warning-container-${inputType}-${action}`}
-    className={`${styles['warning-container']} ${inputType === 'gamepad' ? styles['margin-left'] : ''}`}
-  >
-    {inputType === 'keyboard'
-      ?
-      <Button
-        key={`keyboard-${group}-${action}-${index}`}
-        onClick={() => handleClick(group, action, index, inputType)}
-        className={`${styles.button}`}>
-        {
-          (userConfig?.[group]?.[action]?.keys?.length
-            && parseBindingName(userConfig[group]?.[action]?.keys?.[index]))
-          || (keys?.length && parseBindingName(keys[index]))
-          || ''
-        }
-      </Button>
-      :
-      <Button
-        className={`${styles.button}`}
-        onClick={() => handleClick(group, action, 0, 'gamepad')}
-      >
-        {isPS ? (
-          gamepadButtons?.[0] ? (
-            buttonsMap[gamepadButtons[0]] ? (
-              <div style={{ marginTop: '3px' }} dangerouslySetInnerHTML={{ __html: buttonsMap[gamepadButtons[0]] }}></div>
-            ) : (
-              gamepadButtons[0]
-            )
-          ) : (
-            ''
-          )
-        ) : (
-          gamepadButtons?.[0] ?? ''
-        )}
-      </Button>
-
-    }
-    <div id={`bind-warning-${group}-${action}-${inputType}-${index}`} className={styles['matched-bind-warning']}>
-      <PixelartIcon
-        iconName={'alert'}
-        width={5}
-        styles={{
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          marginRight: '2px'
-        }} />
-      <div>
-        This bind is already in use. <span></span>
-      </div>
-    </div>
-  </div>
-}
-
-export const AwaitingInputOverlay = ({ isGamepad }) => {
-  return <div style={{
-    position: 'fixed',
-    inset: 0,
-    display: 'flex',
-    justifyContent: 'center',
-    alignItems: 'center',
-    backgroundColor: 'rgba(0, 0, 0, 0.5)',
-    color: 'white',
-    fontSize: 24,
-    zIndex: 10
-  }}
-  >
-    {isGamepad ? 'Press the button on the gamepad' : 'Press the key'}.
-    Press ESC to cancel.
-  </div>
-}
-
-const parseActionName = (action: string) => {
-  const parts = action.split(/(?=[A-Z])/)
-  parts[0] = parts[0].charAt(0).toUpperCase() + parts[0].slice(1)
-  return parts.join(' ')
-}
-
-const parseBindingName = (binding: string | undefined) => {
-  if (!binding) return ''
-  const cut = binding.replaceAll(/(Numpad|Digit|Key)/g, '')
-  const parts = cut.split(/(?=[A-Z\d])/)
-  return parts.reverse().join(' ')
-}
-
-const buttonsMap = {
-  'A': `<?xml version="1.0" ?><!DOCTYPE svg  PUBLIC '-//W3C//DTD SVG 1.1//EN'  'http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd'><svg enable-background="new 0 0 15 15" height="15px" id="Livello_1" version="1.1" viewBox="0 0 64 64" width="64px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g><path d="M62,32c0,16.564-13.434,30-30,30C15.433,62,2,48.564,2,32C2,15.433,15.433,2,32,2C48.566,2,62,15.433,62,32   z" fill="#67809F"/><path d="M62,32c0-1.582-0.159-3.122-0.395-4.64c-0.006-0.002-0.011-0.004-0.016-0.009   c-0.02-0.02-0.025-0.045-0.042-0.063c-0.019-0.02-0.045-0.023-0.064-0.043c-0.018-0.019-0.023-0.044-0.042-0.064   c-0.02-0.02-0.047-0.025-0.066-0.044c-0.018-0.019-0.023-0.044-0.042-0.062c-0.019-0.019-0.043-0.023-0.063-0.042   c-0.017-0.021-0.021-0.045-0.041-0.063c-0.019-0.019-0.044-0.024-0.062-0.042c-0.019-0.02-0.024-0.046-0.044-0.066   c-0.02-0.019-0.044-0.025-0.064-0.042c-0.019-0.02-0.024-0.046-0.042-0.065c-0.02-0.019-0.046-0.024-0.064-0.042   c-0.019-0.019-0.024-0.044-0.043-0.062c-0.018-0.021-0.045-0.026-0.064-0.043c-0.017-0.019-0.023-0.043-0.042-0.063   c-0.018-0.018-0.043-0.023-0.063-0.042c-0.019-0.02-0.022-0.045-0.042-0.063c-0.019-0.02-0.044-0.025-0.065-0.043   c-0.017-0.019-0.022-0.044-0.041-0.064c-0.021-0.019-0.046-0.025-0.064-0.042c-0.02-0.018-0.023-0.045-0.043-0.064   c-0.02-0.019-0.044-0.023-0.064-0.042c-0.018-0.021-0.022-0.045-0.041-0.063c-0.019-0.02-0.045-0.024-0.063-0.043   c-0.02-0.018-0.023-0.043-0.044-0.063c-0.016-0.019-0.043-0.023-0.062-0.042c-0.02-0.02-0.025-0.046-0.043-0.065   c-0.021-0.018-0.045-0.024-0.065-0.042c-0.017-0.019-0.021-0.045-0.043-0.063c-0.019-0.021-0.044-0.024-0.063-0.043   c-0.018-0.019-0.022-0.043-0.042-0.063c-0.018-0.018-0.044-0.024-0.063-0.042c-0.02-0.02-0.025-0.045-0.042-0.063   s-0.045-0.023-0.061-0.042c-0.02-0.02-0.026-0.046-0.044-0.065c-0.02-0.019-0.045-0.025-0.065-0.042   c-0.018-0.018-0.022-0.045-0.043-0.064c-0.019-0.019-0.045-0.025-0.065-0.042c-0.016-0.021-0.021-0.045-0.041-0.063   c-0.019-0.02-0.044-0.024-0.063-0.043c-0.017-0.018-0.022-0.043-0.042-0.063c-0.018-0.019-0.044-0.023-0.061-0.042   c-0.021-0.02-0.027-0.046-0.044-0.065c-0.02-0.018-0.044-0.024-0.065-0.042c-0.018-0.019-0.023-0.045-0.042-0.063   c-0.02-0.021-0.045-0.024-0.063-0.043c-0.017-0.019-0.023-0.043-0.043-0.063c-0.019-0.018-0.043-0.023-0.063-0.042   c-0.016-0.02-0.022-0.045-0.042-0.063c-0.018-0.02-0.044-0.025-0.066-0.044c-0.017-0.019-0.022-0.043-0.039-0.062   c-0.021-0.019-0.044-0.025-0.065-0.042c-0.019-0.02-0.022-0.046-0.042-0.065c-0.02-0.019-0.044-0.025-0.063-0.042   c-0.018-0.02-0.024-0.044-0.042-0.063c-0.019-0.018-0.044-0.024-0.063-0.042c-0.018-0.02-0.024-0.046-0.045-0.066   c-0.017-0.019-0.044-0.025-0.064-0.043c-0.018-0.02-0.022-0.044-0.041-0.062c-0.018-0.018-0.043-0.023-0.063-0.042   c-0.019-0.019-0.023-0.045-0.042-0.063c-0.021-0.021-0.045-0.026-0.064-0.043c-0.019-0.019-0.023-0.044-0.042-0.063   c-0.02-0.018-0.044-0.024-0.063-0.041c-0.019-0.02-0.025-0.045-0.044-0.063c-0.018-0.02-0.046-0.024-0.063-0.043   c-0.018-0.018-0.024-0.044-0.042-0.064c-0.019-0.019-0.045-0.024-0.063-0.041c-0.02-0.021-0.024-0.047-0.044-0.066   c-0.018-0.019-0.043-0.025-0.064-0.042c-0.016-0.02-0.022-0.044-0.04-0.063c-0.021-0.02-0.044-0.025-0.064-0.043   c-0.017-0.019-0.023-0.043-0.041-0.063c-0.021-0.019-0.046-0.025-0.063-0.042c-0.021-0.02-0.025-0.046-0.044-0.065   c-0.019-0.018-0.044-0.023-0.063-0.042c-0.019-0.019-0.025-0.045-0.044-0.063c-0.019-0.021-0.045-0.026-0.063-0.043   c-0.019-0.019-0.023-0.044-0.044-0.063c-0.019-0.018-0.046-0.024-0.065-0.043c-0.017-0.02-0.021-0.045-0.041-0.062   c-0.018-0.02-0.043-0.023-0.063-0.043c-0.018-0.018-0.023-0.044-0.043-0.064c-0.02-0.019-0.044-0.025-0.065-0.042   c-0.019-0.02-0.022-0.044-0.041-0.063c-0.019-0.019-0.044-0.023-0.064-0.042c-0.018-0.019-0.022-0.044-0.041-0.062   c-0.018-0.02-0.045-0.024-0.063-0.043c-0.018-0.019-0.024-0.044-0.043-0.064c-0.018-0.019-0.043-0.023-0.063-0.042   c-0.019-0.02-0.024-0.046-0.044-0.065c-0.019-0.018-0.044-0.024-0.066-0.044c-0.017-0.019-0.021-0.044-0.04-0.062   s-0.044-0.023-0.062-0.041c-0.021-0.02-0.024-0.047-0.044-0.066c-0.021-0.018-0.047-0.024-0.065-0.043   c-0.019-0.018-0.023-0.043-0.04-0.06c-0.019-0.019-0.045-0.024-0.063-0.043c-0.019-0.018-0.024-0.044-0.044-0.063   c-0.019-0.02-0.044-0.024-0.063-0.042c-0.018-0.02-0.022-0.045-0.045-0.065c-0.017-0.019-0.043-0.023-0.063-0.042   c-0.018-0.019-0.022-0.044-0.042-0.062c-0.019-0.02-0.044-0.026-0.064-0.043c-0.017-0.019-0.023-0.043-0.04-0.063   c-0.021-0.019-0.045-0.025-0.064-0.042c-0.02-0.02-0.025-0.046-0.043-0.064c-0.02-0.018-0.044-0.023-0.062-0.042   c-0.02-0.021-0.025-0.046-0.044-0.066c-0.02-0.019-0.047-0.024-0.066-0.044c-0.017-0.019-0.023-0.044-0.042-0.062   c-0.019-0.019-0.043-0.023-0.064-0.042c-0.016-0.02-0.021-0.045-0.04-0.063c-0.019-0.019-0.043-0.024-0.063-0.043   c-0.019-0.019-0.024-0.045-0.041-0.064c-0.021-0.02-0.044-0.024-0.065-0.042c-0.019-0.02-0.024-0.045-0.042-0.065   c-0.02-0.019-0.046-0.024-0.064-0.042c-0.019-0.019-0.024-0.044-0.043-0.062c-0.017-0.02-0.045-0.024-0.063-0.043   c-0.018-0.019-0.023-0.043-0.044-0.063c-0.016-0.019-0.043-0.023-0.062-0.04c-0.018-0.021-0.023-0.048-0.043-0.066   c-0.019-0.02-0.043-0.025-0.065-0.043c-0.017-0.019-0.021-0.045-0.041-0.064c-0.021-0.019-0.046-0.024-0.064-0.042   c-0.02-0.018-0.023-0.044-0.043-0.063c-0.018-0.019-0.044-0.023-0.063-0.042c-0.019-0.02-0.023-0.045-0.043-0.063   c-0.017-0.019-0.045-0.024-0.063-0.043s-0.024-0.043-0.043-0.063c-0.016-0.019-0.043-0.023-0.062-0.042   c-0.02-0.02-0.024-0.046-0.043-0.065c-0.021-0.018-0.045-0.024-0.065-0.042c-0.017-0.019-0.021-0.044-0.041-0.062   c-0.021-0.019-0.043-0.024-0.062-0.042c-0.021-0.02-0.025-0.046-0.044-0.066c-0.02-0.019-0.046-0.025-0.065-0.043   c-0.02-0.019-0.025-0.044-0.041-0.062c-0.02-0.02-0.046-0.025-0.065-0.043c-0.018-0.019-0.023-0.045-0.042-0.064   c-0.02-0.018-0.043-0.023-0.062-0.041c-0.019-0.02-0.024-0.047-0.045-0.066c-0.019-0.019-0.043-0.023-0.063-0.042   c-0.016-0.02-0.022-0.045-0.041-0.063c-0.021-0.019-0.045-0.024-0.066-0.044c-0.018-0.019-0.023-0.043-0.04-0.062   c-0.021-0.019-0.046-0.025-0.065-0.042c-0.019-0.02-0.024-0.044-0.04-0.063c-0.02-0.018-0.044-0.023-0.063-0.041   c-0.019-0.021-0.025-0.047-0.044-0.065c-0.018-0.02-0.045-0.024-0.064-0.043c-0.017-0.019-0.023-0.044-0.043-0.064   c-0.018-0.019-0.043-0.023-0.063-0.042c-0.016-0.019-0.021-0.044-0.042-0.063c-0.018-0.02-0.044-0.025-0.066-0.044   c-0.017-0.02-0.021-0.044-0.039-0.062c-0.02-0.019-0.044-0.023-0.063-0.042c-0.021-0.02-0.024-0.046-0.043-0.065   c-0.021-0.019-0.044-0.025-0.063-0.042c-0.019-0.019-0.025-0.044-0.042-0.063c-0.02-0.019-0.046-0.025-0.063-0.042   c-0.018-0.019-0.024-0.045-0.046-0.064c-0.016-0.02-0.043-0.025-0.063-0.042c-0.017-0.02-0.022-0.044-0.041-0.063   c-0.019-0.018-0.043-0.023-0.063-0.042c-0.019-0.02-0.022-0.044-0.042-0.063c-0.019-0.019-0.044-0.023-0.063-0.042   c-0.021-0.018-0.025-0.045-0.042-0.064c-0.021-0.019-0.046-0.025-0.064-0.042c-0.019-0.02-0.025-0.046-0.044-0.065   c-0.018-0.02-0.046-0.025-0.065-0.043c-0.018-0.019-0.023-0.043-0.04-0.063c-0.019-0.018-0.045-0.023-0.063-0.041   c-0.018-0.021-0.024-0.047-0.045-0.066c-0.018-0.019-0.043-0.025-0.064-0.043c-0.016-0.019-0.021-0.044-0.04-0.063   c-0.021-0.019-0.046-0.025-0.067-0.044c-0.018-0.019-0.021-0.043-0.04-0.062c-0.019-0.019-0.043-0.023-0.062-0.041   c-0.02-0.021-0.025-0.047-0.044-0.066c-0.019-0.018-0.044-0.023-0.063-0.042c-0.018-0.02-0.024-0.044-0.043-0.063   c-0.017-0.019-0.045-0.023-0.062-0.042c-0.019-0.018-0.024-0.045-0.045-0.064c-0.019-0.019-0.046-0.025-0.065-0.043   c-0.017-0.019-0.021-0.044-0.041-0.062c-0.018-0.018-0.041-0.023-0.062-0.042c-0.019-0.02-0.024-0.046-0.043-0.066   c-0.021-0.019-0.045-0.025-0.064-0.042c-0.019-0.02-0.024-0.044-0.043-0.063c-0.018-0.019-0.044-0.023-0.061-0.042   c-0.021-0.019-0.026-0.045-0.045-0.064c-0.02-0.019-0.048-0.025-0.065-0.044c-0.018-0.019-0.022-0.043-0.043-0.062   c-0.017-0.019-0.043-0.024-0.061-0.042c-0.019-0.019-0.023-0.044-0.044-0.063c-0.019-0.018-0.041-0.023-0.062-0.042   c-0.019-0.021-0.023-0.046-0.043-0.066c-0.021-0.021-0.046-0.026-0.068-0.043c-0.017-0.019-0.021-0.044-0.038-0.062   c-0.021-0.018-0.045-0.024-0.062-0.042c-0.021-0.021-0.026-0.046-0.044-0.065c-0.02-0.02-0.046-0.023-0.064-0.043   c-0.019-0.019-0.024-0.045-0.044-0.064c-0.019-0.019-0.044-0.023-0.063-0.042c-0.017-0.02-0.022-0.044-0.043-0.063   c-0.019-0.019-0.045-0.025-0.066-0.044c-0.016-0.019-0.021-0.043-0.04-0.061c-0.019-0.02-0.043-0.024-0.063-0.043   c-0.018-0.019-0.024-0.045-0.041-0.064c-0.021-0.02-0.044-0.025-0.064-0.042c-0.02-0.019-0.025-0.044-0.041-0.063   c-0.02-0.02-0.046-0.023-0.064-0.042c-0.019-0.02-0.024-0.044-0.041-0.063c-0.021-0.019-0.049-0.024-0.066-0.044   c-0.019-0.018-0.023-0.044-0.044-0.063c-0.018-0.019-0.043-0.024-0.063-0.042c-0.017-0.02-0.021-0.044-0.041-0.063   c-0.019-0.019-0.043-0.024-0.063-0.043c-0.019-0.019-0.023-0.045-0.041-0.064c-0.021-0.018-0.044-0.024-0.062-0.041   c-0.022-0.021-0.027-0.047-0.044-0.066c-0.021-0.019-0.046-0.025-0.064-0.043c-0.019-0.019-0.024-0.044-0.043-0.062   c-0.019-0.02-0.047-0.026-0.065-0.044c-0.018-0.019-0.024-0.043-0.043-0.062c-0.018-0.019-0.043-0.023-0.063-0.042   c-0.018-0.019-0.021-0.044-0.042-0.063c-0.019-0.02-0.043-0.025-0.065-0.042c-0.017-0.02-0.021-0.045-0.041-0.065   c-0.021-0.019-0.042-0.023-0.062-0.041c-0.021-0.02-0.024-0.047-0.043-0.066c-0.021-0.019-0.047-0.025-0.065-0.042   c-0.019-0.02-0.024-0.044-0.042-0.063c-0.879-0.879-2.303-0.879-3.182,0L32.001,28.819L18.69,15.506   c-0.002,0-0.003-0.001-0.003-0.001l-0.104-0.104c-0.001,0-0.001-0.002-0.001-0.002l-0.531-0.529c-0.879-0.879-2.302-0.879-3.181,0   s-0.879,2.303,0,3.182l0.317,0.316c0.002,0.001,0.002,0.002,0.002,0.002l0.317,0.318l0.425,0.425l0.53,0.53l0.638,0.637   l0.105,0.105l1.167,1.168l0.107,0.106l0.53,0.531l0.636,0.635v0.001l9.108,9.106L28.818,32L14.87,45.948   c-0.879,0.879-0.879,2.303,0,3.182c0.019,0.018,0.043,0.022,0.064,0.043c0.018,0.019,0.023,0.044,0.042,0.064   c0.019,0.019,0.044,0.022,0.063,0.042c0.018,0.019,0.023,0.043,0.042,0.063c0.02,0.02,0.046,0.025,0.065,0.042   c0.019,0.021,0.022,0.046,0.042,0.064c0.019,0.021,0.046,0.025,0.064,0.042c0.018,0.02,0.023,0.046,0.041,0.062   c0.02,0.021,0.044,0.026,0.065,0.044c0.017,0.02,0.022,0.046,0.042,0.064c0.019,0.017,0.042,0.022,0.061,0.041   c0.02,0.021,0.025,0.047,0.044,0.065c0.02,0.018,0.046,0.024,0.065,0.043c0.017,0.02,0.022,0.043,0.042,0.063   c0.018,0.019,0.044,0.024,0.064,0.041c0.018,0.021,0.024,0.045,0.042,0.064c0.019,0.021,0.043,0.024,0.064,0.042   c0.018,0.021,0.023,0.046,0.041,0.063c0.019,0.02,0.044,0.024,0.065,0.043c0.018,0.019,0.023,0.047,0.042,0.064   c0.02,0.019,0.046,0.025,0.065,0.043c0.018,0.018,0.023,0.044,0.042,0.063c0.017,0.017,0.044,0.023,0.063,0.042   c0.017,0.02,0.022,0.043,0.042,0.063c0.018,0.019,0.044,0.024,0.065,0.041c0.017,0.021,0.022,0.045,0.041,0.065   c0.021,0.019,0.045,0.025,0.064,0.041c0.018,0.021,0.023,0.045,0.042,0.063c0.02,0.021,0.044,0.024,0.063,0.043   c0.017,0.019,0.023,0.045,0.042,0.063c0.02,0.019,0.044,0.023,0.063,0.042c0.018,0.018,0.024,0.046,0.044,0.065   c0.017,0.018,0.043,0.021,0.062,0.042c0.019,0.019,0.025,0.045,0.043,0.065c0.019,0.018,0.046,0.022,0.065,0.041   c0.018,0.02,0.023,0.043,0.041,0.063c0.021,0.02,0.045,0.024,0.064,0.041c0.018,0.021,0.023,0.045,0.042,0.063   c0.019,0.021,0.044,0.025,0.063,0.043c0.017,0.019,0.023,0.046,0.043,0.063c0.018,0.021,0.044,0.024,0.062,0.044   c0.019,0.018,0.025,0.046,0.044,0.064c0.019,0.018,0.043,0.022,0.064,0.043c0.018,0.019,0.022,0.043,0.041,0.063   c0.019,0.018,0.044,0.023,0.064,0.042c0.017,0.02,0.023,0.045,0.042,0.065c0.02,0.019,0.046,0.024,0.065,0.041   c0.018,0.021,0.023,0.045,0.042,0.063c0.017,0.021,0.043,0.025,0.063,0.041c0.017,0.021,0.023,0.046,0.042,0.063   c0.018,0.021,0.044,0.026,0.065,0.044c0.018,0.018,0.023,0.046,0.041,0.064c0.019,0.019,0.044,0.022,0.064,0.042   c0.018,0.02,0.023,0.044,0.041,0.063c0.02,0.018,0.047,0.023,0.064,0.042c0.019,0.02,0.023,0.043,0.042,0.063   c0.019,0.019,0.045,0.024,0.064,0.042c0.019,0.02,0.023,0.044,0.042,0.064c0.019,0.02,0.043,0.025,0.063,0.041   c0.02,0.021,0.024,0.046,0.043,0.063c0.018,0.021,0.044,0.025,0.063,0.044c0.019,0.018,0.025,0.045,0.043,0.064   c0.02,0.019,0.044,0.023,0.064,0.042c0.018,0.018,0.024,0.046,0.043,0.065c0.018,0.018,0.044,0.023,0.065,0.042   c0.016,0.019,0.021,0.043,0.04,0.062s0.045,0.023,0.064,0.042c0.018,0.021,0.024,0.044,0.043,0.064   c0.018,0.02,0.044,0.024,0.063,0.042c0.02,0.02,0.025,0.046,0.043,0.064c0.019,0.021,0.044,0.025,0.064,0.043   c0.017,0.019,0.023,0.044,0.041,0.062c0.021,0.021,0.046,0.026,0.064,0.045c0.019,0.017,0.024,0.045,0.044,0.063   c0.018,0.019,0.043,0.024,0.063,0.043c0.018,0.019,0.024,0.043,0.043,0.063c0.018,0.02,0.044,0.022,0.062,0.043   c0.019,0.019,0.024,0.044,0.044,0.064c0.02,0.02,0.047,0.024,0.066,0.043c0.018,0.021,0.022,0.045,0.042,0.062   c0.018,0.021,0.043,0.024,0.063,0.042c0.018,0.02,0.023,0.046,0.042,0.063c0.019,0.021,0.044,0.025,0.063,0.045   c0.018,0.018,0.023,0.045,0.044,0.063c0.018,0.019,0.044,0.024,0.064,0.043c0.018,0.019,0.023,0.044,0.041,0.063   c0.018,0.017,0.044,0.022,0.062,0.041c0.019,0.021,0.023,0.046,0.044,0.066c0.02,0.019,0.046,0.024,0.064,0.042   c0.018,0.02,0.023,0.043,0.042,0.063c0.021,0.02,0.045,0.026,0.064,0.042c0.018,0.021,0.023,0.046,0.044,0.064   c0.018,0.021,0.044,0.026,0.063,0.044c0.018,0.018,0.023,0.044,0.043,0.063c0.018,0.019,0.044,0.023,0.061,0.043   c0.02,0.018,0.025,0.045,0.044,0.064c0.02,0.018,0.046,0.023,0.066,0.044c0.017,0.019,0.021,0.043,0.041,0.062   s0.044,0.023,0.063,0.042c0.018,0.02,0.023,0.043,0.043,0.063c0.019,0.018,0.041,0.023,0.061,0.039   c0.02,0.022,0.025,0.049,0.044,0.066c0.019,0.021,0.046,0.026,0.065,0.044c0.017,0.019,0.022,0.045,0.042,0.063   c0.02,0.021,0.046,0.026,0.064,0.045c0.019,0.017,0.024,0.044,0.042,0.062c0.021,0.019,0.045,0.025,0.064,0.044   c0.018,0.019,0.023,0.043,0.042,0.063c0.02,0.018,0.043,0.021,0.062,0.04c0.019,0.021,0.024,0.047,0.044,0.066   s0.046,0.025,0.065,0.042c0.018,0.021,0.023,0.045,0.042,0.063c0.017,0.02,0.043,0.025,0.064,0.043   c0.018,0.019,0.023,0.046,0.041,0.063c0.021,0.021,0.047,0.027,0.065,0.044c0.019,0.019,0.023,0.045,0.043,0.063   c0.019,0.018,0.043,0.022,0.063,0.042c0.018,0.02,0.023,0.044,0.042,0.063c0.019,0.018,0.046,0.023,0.065,0.042   c0.017,0.02,0.022,0.045,0.042,0.065c0.018,0.017,0.044,0.022,0.062,0.039c0.02,0.022,0.024,0.047,0.044,0.066   c0.019,0.02,0.045,0.026,0.064,0.042c0.018,0.021,0.023,0.045,0.04,0.063c0.021,0.02,0.047,0.026,0.065,0.043   c0.017,0.019,0.022,0.044,0.042,0.063c0.02,0.019,0.044,0.022,0.063,0.042c0.018,0.02,0.024,0.046,0.044,0.065   c0.017,0.017,0.043,0.021,0.062,0.042c0.019,0.019,0.024,0.045,0.043,0.064c0.02,0.019,0.046,0.023,0.065,0.042   c0.018,0.02,0.023,0.043,0.041,0.063c0.019,0.019,0.043,0.023,0.064,0.041c0.018,0.021,0.023,0.045,0.041,0.063   c0.02,0.021,0.045,0.025,0.064,0.043c0.017,0.019,0.023,0.044,0.042,0.063c0.02,0.019,0.046,0.024,0.064,0.042   s0.023,0.046,0.043,0.064c0.017,0.019,0.043,0.022,0.062,0.042c0.019,0.02,0.025,0.046,0.043,0.065   c0.02,0.019,0.045,0.021,0.063,0.042c0.019,0.02,0.023,0.043,0.043,0.063c0.021,0.02,0.047,0.024,0.066,0.043   c0.018,0.021,0.022,0.045,0.041,0.062c0.019,0.021,0.044,0.024,0.063,0.043c0.017,0.019,0.023,0.046,0.043,0.063   c0.018,0.021,0.044,0.026,0.063,0.044c0.018,0.018,0.024,0.044,0.043,0.063c0.018,0.019,0.043,0.023,0.062,0.042   c0.019,0.02,0.025,0.046,0.044,0.065c0.02,0.018,0.047,0.023,0.066,0.044c0.017,0.02,0.021,0.043,0.04,0.062   c0.02,0.019,0.046,0.024,0.064,0.041c0.019,0.021,0.023,0.047,0.043,0.065c0.019,0.02,0.043,0.025,0.063,0.041   c0.019,0.021,0.023,0.046,0.041,0.063c0.02,0.021,0.045,0.026,0.063,0.043c0.019,0.019,0.023,0.044,0.043,0.063   c0.019,0.019,0.044,0.024,0.064,0.043c0.018,0.019,0.024,0.045,0.042,0.064c0.019,0.018,0.044,0.022,0.063,0.042   c0.017,0.02,0.023,0.045,0.043,0.065c0.019,0.019,0.045,0.024,0.063,0.041c0.018,0.02,0.022,0.043,0.043,0.063   c0.018,0.02,0.043,0.025,0.064,0.041c0.018,0.021,0.023,0.045,0.041,0.063c0.018,0.021,0.044,0.024,0.063,0.044   c0.019,0.018,0.025,0.045,0.043,0.064c0.02,0.019,0.044,0.023,0.063,0.042c0.019,0.018,0.024,0.046,0.044,0.064   c0.018,0.019,0.044,0.024,0.064,0.043c0.017,0.019,0.023,0.043,0.041,0.063c0.019,0.02,0.045,0.023,0.063,0.042   c0.017,0.02,0.023,0.043,0.043,0.063c0.018,0.02,0.044,0.023,0.062,0.042c0.02,0.02,0.025,0.046,0.044,0.064   c0.02,0.02,0.046,0.025,0.066,0.045c0.017,0.019,0.021,0.044,0.04,0.06c0.02,0.019,0.043,0.025,0.062,0.043   c0.02,0.019,0.025,0.047,0.045,0.065c0.019,0.019,0.043,0.023,0.063,0.042c0.018,0.02,0.023,0.044,0.042,0.063   c0.02,0.02,0.046,0.023,0.066,0.044c0.017,0.02,0.023,0.043,0.041,0.062c0.02,0.02,0.044,0.025,0.063,0.042   c0.019,0.021,0.024,0.046,0.044,0.064c0.018,0.02,0.043,0.024,0.063,0.042c0.018,0.02,0.023,0.046,0.042,0.063   c0.02,0.02,0.043,0.025,0.062,0.042c0.02,0.021,0.024,0.047,0.045,0.066c0.019,0.019,0.045,0.024,0.063,0.043   c0.019,0.019,0.025,0.045,0.043,0.064c0.019,0.018,0.044,0.022,0.063,0.042c0.018,0.019,0.022,0.043,0.042,0.063   c0.021,0.019,0.047,0.024,0.066,0.043c0.018,0.02,0.022,0.043,0.042,0.062c0.019,0.02,0.043,0.024,0.063,0.042   c0.018,0.021,0.023,0.045,0.042,0.063c0.019,0.02,0.046,0.025,0.064,0.043c0.018,0.019,0.023,0.046,0.042,0.064   c0.018,0.019,0.044,0.023,0.063,0.043c0.02,0.018,0.025,0.045,0.043,0.063c0.019,0.019,0.043,0.023,0.064,0.044   c0.018,0.018,0.023,0.043,0.04,0.062c0.021,0.02,0.048,0.024,0.065,0.043c0.018,0.02,0.022,0.043,0.043,0.063   c0.019,0.019,0.043,0.022,0.062,0.039c0.018,0.022,0.024,0.049,0.044,0.066c0.017,0.02,0.043,0.024,0.062,0.044   c0.019,0.019,0.024,0.045,0.043,0.063c0.021,0.021,0.047,0.026,0.064,0.044c0.019,0.018,0.023,0.045,0.042,0.063   c0.021,0.019,0.045,0.025,0.064,0.043c0.018,0.02,0.023,0.044,0.042,0.063c0.019,0.018,0.043,0.021,0.062,0.04   c0.019,0.021,0.024,0.047,0.044,0.066c0.02,0.019,0.046,0.025,0.064,0.042c0.019,0.02,0.023,0.046,0.043,0.064   c0.019,0.02,0.043,0.024,0.063,0.042c0.019,0.02,0.022,0.045,0.041,0.063c0.021,0.021,0.045,0.025,0.063,0.044   c0.019,0.019,0.023,0.044,0.043,0.063c0.021,0.018,0.047,0.024,0.065,0.044c0.019,0.019,0.023,0.044,0.042,0.062   s0.045,0.023,0.065,0.042c0.017,0.02,0.022,0.045,0.041,0.064c0.019,0.018,0.045,0.023,0.062,0.04   c0.021,0.021,0.024,0.047,0.045,0.066c0.019,0.02,0.043,0.024,0.063,0.042c0.019,0.021,0.023,0.045,0.041,0.063   c0.021,0.02,0.048,0.025,0.066,0.045c0.018,0.018,0.021,0.044,0.041,0.062c0.02,0.019,0.044,0.022,0.063,0.042   s0.024,0.046,0.045,0.064c0.017,0.018,0.043,0.024,0.064,0.043c0.003,0.004,0.005,0.01,0.008,0.016C28.879,61.841,30.419,62,32,62   C48.566,62,62,48.564,62,32z" opacity="0.2"/><path d="M35.183,32L49.13,18.052c0.879-0.879,0.879-2.303,0-3.182s-2.303-0.879-3.182,0L32,28.819L18.051,14.871   c-0.879-0.879-2.302-0.879-3.181,0s-0.879,2.303,0,3.182L28.818,32L14.87,45.948c-0.879,0.879-0.879,2.303,0,3.182   c0.439,0.438,1.017,0.657,1.591,0.657c0.577,0,1.151-0.219,1.59-0.657L32,35.183L45.948,49.13c0.438,0.438,1.016,0.657,1.591,0.657   c0.577,0,1.15-0.219,1.591-0.657c0.879-0.879,0.879-2.303,0-3.182L35.183,32z" fill="#FFFFFF"/></g></svg>`,
-  'B': `<?xml version="1.0" ?><!DOCTYPE svg  PUBLIC '-//W3C//DTD SVG 1.1//EN'  'http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd'><svg enable-background="new 0 0 15 15" height="15px" id="Livello_1" version="1.1" viewBox="0 0 64 64" width="64px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g><path d="M62,32c0,16.564-13.436,30-30,30C15.433,62,2,48.564,2,32C2,15.433,15.433,2,32,2C48.564,2,62,15.433,62,32   z" fill="#D24D57"/><path d="M61.867,34.598c-0.012-0.012-0.021-0.023-0.031-0.035c-0.039-0.035-0.072-0.072-0.108-0.108   c-0.036-0.034-0.069-0.071-0.108-0.104c-0.033-0.035-0.067-0.071-0.104-0.105c-0.035-0.038-0.072-0.073-0.107-0.107   c-0.036-0.037-0.07-0.068-0.102-0.105c-0.039-0.036-0.076-0.071-0.111-0.111c-0.035-0.029-0.064-0.063-0.1-0.094   c-0.035-0.039-0.073-0.078-0.113-0.113c-0.035-0.039-0.069-0.073-0.107-0.107c-0.033-0.038-0.07-0.072-0.105-0.109   c-0.034-0.033-0.07-0.066-0.104-0.104c-0.037-0.035-0.072-0.073-0.106-0.107c-0.036-0.032-0.067-0.068-0.104-0.1   c-0.035-0.04-0.071-0.075-0.111-0.111c-0.033-0.037-0.07-0.07-0.104-0.106c-0.032-0.033-0.066-0.067-0.099-0.101   c-0.042-0.039-0.081-0.08-0.122-0.12c-0.033-0.03-0.063-0.063-0.098-0.095c-0.031-0.034-0.065-0.069-0.102-0.104   c-0.037-0.037-0.076-0.076-0.115-0.115c-0.033-0.036-0.07-0.069-0.105-0.105c-0.032-0.033-0.066-0.066-0.099-0.1   c-0.038-0.037-0.071-0.072-0.108-0.109c-0.037-0.036-0.071-0.071-0.109-0.105c-0.035-0.038-0.071-0.075-0.109-0.112   c-0.031-0.032-0.066-0.063-0.098-0.097c-0.035-0.036-0.07-0.073-0.106-0.108c-0.038-0.036-0.075-0.074-0.111-0.111   c-0.036-0.034-0.067-0.068-0.103-0.101c-0.035-0.037-0.071-0.075-0.111-0.112c-0.033-0.032-0.07-0.067-0.102-0.103   c-0.036-0.036-0.072-0.072-0.107-0.106c-0.039-0.036-0.072-0.071-0.106-0.106c-0.036-0.036-0.069-0.07-0.108-0.105   c-0.035-0.037-0.071-0.074-0.109-0.108c-0.031-0.034-0.066-0.067-0.1-0.102c-0.034-0.037-0.072-0.074-0.107-0.108   c-0.035-0.037-0.07-0.072-0.105-0.107c-0.039-0.035-0.072-0.07-0.108-0.106c-0.034-0.036-0.067-0.07-0.106-0.104   c-0.035-0.038-0.071-0.075-0.109-0.112c-0.031-0.033-0.066-0.064-0.098-0.098c-0.035-0.036-0.07-0.071-0.104-0.106   c-0.04-0.038-0.077-0.078-0.117-0.116c-0.036-0.036-0.069-0.068-0.106-0.104c-0.031-0.033-0.063-0.066-0.098-0.1   c-0.035-0.037-0.072-0.072-0.107-0.108c-0.036-0.037-0.072-0.073-0.107-0.107c-0.039-0.037-0.074-0.073-0.109-0.109   c-0.035-0.034-0.066-0.066-0.104-0.102c-0.035-0.036-0.069-0.072-0.107-0.108c-0.033-0.035-0.07-0.07-0.105-0.106   c-0.035-0.037-0.074-0.071-0.109-0.107c-0.031-0.034-0.066-0.067-0.099-0.101c-0.038-0.037-0.073-0.072-0.109-0.109   s-0.071-0.073-0.11-0.108c-0.033-0.036-0.069-0.071-0.107-0.107c-0.031-0.034-0.066-0.066-0.1-0.101   c-0.034-0.036-0.072-0.073-0.105-0.109c-0.039-0.035-0.072-0.071-0.108-0.107c-0.038-0.037-0.071-0.07-0.11-0.106   c-0.031-0.035-0.065-0.069-0.104-0.105c-0.031-0.034-0.066-0.067-0.1-0.102c-0.038-0.041-0.08-0.079-0.119-0.118   c-0.032-0.034-0.066-0.068-0.102-0.102c-0.033-0.033-0.066-0.066-0.102-0.099c-0.037-0.039-0.073-0.076-0.111-0.112   c-0.033-0.035-0.068-0.069-0.102-0.104c-0.037-0.038-0.076-0.074-0.111-0.111c-0.034-0.034-0.066-0.067-0.101-0.102   c-0.04-0.036-0.075-0.073-0.111-0.111c-0.032-0.032-0.063-0.064-0.097-0.096c-0.037-0.038-0.077-0.078-0.117-0.117   c-0.031-0.033-0.066-0.064-0.098-0.098c-0.038-0.039-0.078-0.077-0.113-0.113c-0.035-0.034-0.066-0.067-0.101-0.101   c-0.04-0.039-0.075-0.075-0.114-0.112c-0.031-0.033-0.063-0.066-0.098-0.099c-0.035-0.037-0.076-0.075-0.113-0.112   c-0.033-0.036-0.068-0.07-0.102-0.103c-0.039-0.037-0.074-0.073-0.107-0.108c-0.037-0.035-0.068-0.068-0.107-0.103   c-0.033-0.037-0.067-0.072-0.105-0.108c-0.033-0.035-0.07-0.071-0.107-0.107c-0.033-0.036-0.07-0.069-0.104-0.104   c-0.035-0.036-0.07-0.071-0.105-0.107c-0.039-0.034-0.072-0.071-0.108-0.107s-0.069-0.071-0.106-0.104   c-0.035-0.037-0.069-0.071-0.105-0.107c-0.033-0.035-0.07-0.067-0.102-0.102c-0.038-0.039-0.078-0.077-0.113-0.113   c-0.035-0.034-0.066-0.067-0.101-0.102c-0.038-0.038-0.073-0.072-0.114-0.109c-0.031-0.035-0.065-0.07-0.102-0.105   c-0.035-0.034-0.072-0.07-0.105-0.105c-0.036-0.037-0.074-0.073-0.109-0.108s-0.068-0.069-0.101-0.104   c-0.038-0.036-0.071-0.071-0.11-0.107c-0.035-0.037-0.068-0.072-0.105-0.107c-0.035-0.036-0.073-0.073-0.111-0.11   c-0.029-0.033-0.066-0.066-0.096-0.098c-0.039-0.037-0.074-0.075-0.111-0.113c-0.037-0.033-0.07-0.069-0.104-0.104   c-0.038-0.038-0.075-0.074-0.114-0.111c-0.029-0.032-0.062-0.065-0.096-0.098c-0.035-0.036-0.072-0.071-0.105-0.105   c-0.038-0.04-0.078-0.078-0.113-0.115c-0.039-0.036-0.072-0.069-0.105-0.105c-0.035-0.035-0.068-0.068-0.104-0.102   c-0.035-0.037-0.069-0.073-0.107-0.108c-0.035-0.034-0.072-0.071-0.105-0.106c-0.035-0.037-0.076-0.074-0.111-0.109   c-0.033-0.033-0.066-0.066-0.097-0.1c-0.038-0.038-0.073-0.073-0.112-0.111c-0.035-0.034-0.066-0.068-0.102-0.101   c-0.037-0.038-0.075-0.076-0.113-0.115c-0.033-0.034-0.068-0.066-0.102-0.1c-0.036-0.037-0.074-0.075-0.111-0.112   c-0.033-0.032-0.066-0.065-0.1-0.1c-0.037-0.037-0.072-0.074-0.113-0.11c-0.031-0.034-0.063-0.067-0.098-0.101   c-0.037-0.039-0.078-0.076-0.115-0.114c-0.032-0.034-0.066-0.067-0.098-0.101c-0.041-0.037-0.076-0.074-0.113-0.111   c-0.033-0.033-0.066-0.066-0.1-0.098c-0.035-0.036-0.069-0.072-0.107-0.108c-0.035-0.036-0.072-0.072-0.107-0.109   s-0.074-0.073-0.109-0.109c-0.035-0.034-0.068-0.067-0.101-0.101c-0.04-0.038-0.075-0.074-0.113-0.112   c-0.032-0.033-0.065-0.066-0.099-0.098c-0.035-0.038-0.074-0.075-0.113-0.112c-0.031-0.034-0.068-0.068-0.1-0.101   c-0.037-0.038-0.076-0.076-0.111-0.113c-0.035-0.034-0.07-0.067-0.103-0.102c-0.038-0.037-0.071-0.072-0.11-0.107   c-0.033-0.035-0.067-0.069-0.102-0.104c-0.035-0.036-0.072-0.071-0.107-0.107c-0.036-0.037-0.072-0.072-0.107-0.107   c-0.039-0.038-0.076-0.075-0.111-0.111c-0.033-0.033-0.066-0.064-0.1-0.098c-0.035-0.038-0.071-0.075-0.111-0.111   c-0.033-0.035-0.07-0.07-0.105-0.105c-0.031-0.035-0.066-0.068-0.1-0.102c-0.039-0.036-0.074-0.074-0.111-0.111   c-0.037-0.035-0.068-0.069-0.105-0.104c-0.035-0.036-0.07-0.071-0.107-0.106c-0.033-0.035-0.068-0.069-0.104-0.105   c-0.035-0.037-0.074-0.073-0.111-0.111c-0.032-0.032-0.064-0.065-0.096-0.097c-0.041-0.039-0.078-0.076-0.114-0.115   c-0.036-0.035-0.067-0.068-0.104-0.102c-3.494-3.523-8.334-5.707-13.69-5.707c-10.651,0-19.29,8.636-19.29,19.288   c0,5.352,2.182,10.189,5.702,13.683c0.039,0.041,0.076,0.077,0.114,0.116c0.036,0.035,0.071,0.068,0.105,0.105   c0.034,0.032,0.068,0.067,0.102,0.1c0.035,0.035,0.071,0.074,0.107,0.108c0.036,0.037,0.071,0.072,0.106,0.105   c0.035,0.038,0.069,0.071,0.105,0.107c0.035,0.036,0.07,0.069,0.106,0.106c0.037,0.036,0.074,0.071,0.112,0.111   c0.031,0.031,0.065,0.066,0.097,0.098c0.039,0.037,0.079,0.078,0.116,0.115c0.035,0.036,0.069,0.069,0.104,0.102   c0.034,0.036,0.067,0.068,0.101,0.104c0.035,0.033,0.07,0.067,0.106,0.106c0.034,0.032,0.069,0.07,0.104,0.104   c0.037,0.036,0.074,0.075,0.109,0.11c0.037,0.037,0.072,0.072,0.109,0.106c0.033,0.035,0.065,0.067,0.099,0.103   c0.039,0.037,0.076,0.073,0.113,0.113c0.034,0.032,0.068,0.066,0.102,0.1c0.037,0.035,0.073,0.074,0.109,0.109   c0.034,0.036,0.069,0.07,0.104,0.103c0.036,0.039,0.072,0.074,0.109,0.109c0.033,0.035,0.066,0.067,0.099,0.104   c0.039,0.037,0.077,0.072,0.115,0.113c0.033,0.032,0.067,0.068,0.1,0.101c0.037,0.034,0.073,0.073,0.109,0.108   c0.034,0.035,0.067,0.067,0.101,0.102c0.039,0.039,0.076,0.075,0.114,0.115c0.034,0.032,0.066,0.064,0.101,0.1   c0.036,0.036,0.073,0.074,0.112,0.111c0.032,0.031,0.066,0.066,0.097,0.098c0.039,0.04,0.077,0.076,0.115,0.113   c0.031,0.035,0.065,0.066,0.097,0.101c0.037,0.037,0.076,0.073,0.112,0.112c0.036,0.034,0.071,0.07,0.108,0.105   c0.034,0.033,0.067,0.069,0.101,0.102c0.035,0.036,0.07,0.071,0.106,0.105c0.036,0.039,0.073,0.075,0.11,0.11   c0.036,0.038,0.07,0.069,0.104,0.107c0.035,0.033,0.071,0.067,0.106,0.104c0.035,0.034,0.067,0.07,0.104,0.104   c0.037,0.035,0.074,0.074,0.11,0.109c0.034,0.035,0.068,0.069,0.103,0.102c0.037,0.039,0.073,0.074,0.11,0.112   c0.037,0.036,0.073,0.071,0.109,0.109c0.032,0.031,0.064,0.064,0.098,0.096c0.035,0.035,0.072,0.074,0.107,0.108   c0.037,0.038,0.073,0.074,0.11,0.109c0.036,0.039,0.073,0.075,0.112,0.111c0.031,0.033,0.062,0.061,0.093,0.094   c0.034,0.035,0.069,0.072,0.106,0.107c0.04,0.039,0.076,0.08,0.117,0.117c0.033,0.033,0.069,0.07,0.103,0.103   c0.034,0.036,0.068,0.068,0.102,0.104c0.035,0.034,0.069,0.067,0.103,0.104c0.039,0.037,0.077,0.073,0.115,0.114   c0.033,0.032,0.067,0.067,0.102,0.101c0.035,0.034,0.071,0.072,0.105,0.105c0.037,0.038,0.074,0.074,0.113,0.109   c0.032,0.035,0.064,0.067,0.097,0.102c0.037,0.036,0.074,0.07,0.111,0.109c0.034,0.034,0.07,0.071,0.107,0.106   c0.035,0.034,0.07,0.071,0.106,0.105c0.033,0.036,0.069,0.071,0.104,0.104c0.035,0.037,0.069,0.07,0.104,0.106   c0.036,0.036,0.073,0.071,0.107,0.109c0.039,0.036,0.076,0.071,0.113,0.11c0.033,0.031,0.064,0.066,0.099,0.097   c0.037,0.039,0.074,0.075,0.111,0.111c0.034,0.036,0.068,0.068,0.102,0.103c0.035,0.035,0.07,0.068,0.104,0.106   c0.037,0.036,0.075,0.071,0.112,0.111c0.035,0.034,0.071,0.072,0.108,0.106c0.032,0.032,0.066,0.066,0.098,0.098   c0.036,0.039,0.073,0.074,0.11,0.109c0.035,0.038,0.069,0.072,0.104,0.108c0.037,0.035,0.074,0.07,0.11,0.108   c0.034,0.033,0.068,0.069,0.102,0.102c0.037,0.035,0.073,0.074,0.11,0.109c0.032,0.035,0.066,0.068,0.101,0.101   c0.038,0.04,0.075,0.077,0.114,0.114c0.032,0.034,0.066,0.065,0.098,0.101c0.038,0.036,0.076,0.075,0.115,0.113   c0.032,0.031,0.064,0.066,0.097,0.098c0.039,0.039,0.076,0.075,0.112,0.111c0.034,0.036,0.068,0.069,0.104,0.104   c0.037,0.037,0.072,0.072,0.111,0.11c0.032,0.033,0.066,0.065,0.099,0.1c0.038,0.036,0.074,0.074,0.111,0.11   c0.035,0.033,0.069,0.07,0.103,0.103c0.036,0.037,0.073,0.073,0.11,0.108c0.035,0.039,0.07,0.073,0.107,0.111   c0.034,0.033,0.068,0.064,0.102,0.1c0.034,0.035,0.069,0.071,0.105,0.105c0.037,0.035,0.073,0.074,0.109,0.109   c0.037,0.038,0.074,0.073,0.109,0.107c0.033,0.036,0.067,0.068,0.1,0.103c0.034,0.034,0.07,0.067,0.105,0.105   c0.038,0.036,0.076,0.075,0.115,0.114c0.033,0.033,0.069,0.069,0.104,0.104c0.034,0.033,0.068,0.068,0.102,0.1   c0.032,0.037,0.068,0.07,0.104,0.107c0.039,0.037,0.075,0.074,0.114,0.114c0.034,0.032,0.067,0.064,0.102,0.101   c0.035,0.035,0.071,0.072,0.108,0.107s0.072,0.073,0.107,0.106c0.034,0.036,0.068,0.069,0.102,0.101   c0.037,0.04,0.072,0.075,0.111,0.113c0.034,0.034,0.068,0.066,0.102,0.102c0.037,0.038,0.076,0.077,0.116,0.115   c0.03,0.029,0.061,0.062,0.091,0.092c0.038,0.039,0.075,0.077,0.115,0.113c0.032,0.037,0.068,0.07,0.104,0.106   c0.035,0.035,0.072,0.069,0.107,0.108c0.034,0.033,0.07,0.069,0.105,0.104s0.07,0.072,0.107,0.107   c0.034,0.035,0.07,0.07,0.105,0.104c0.033,0.037,0.067,0.069,0.102,0.105c0.039,0.037,0.075,0.074,0.114,0.114   c0.035,0.033,0.069,0.066,0.104,0.103c0.032,0.033,0.068,0.069,0.104,0.104c0.034,0.032,0.066,0.067,0.101,0.099   c0.038,0.042,0.078,0.082,0.12,0.12c0.033,0.035,0.067,0.067,0.1,0.102c0.034,0.034,0.07,0.067,0.105,0.105   c0.034,0.032,0.067,0.067,0.102,0.102c0.039,0.036,0.075,0.077,0.114,0.114c0.035,0.035,0.071,0.07,0.106,0.104   c0.033,0.036,0.067,0.068,0.1,0.103c0.037,0.036,0.074,0.071,0.109,0.109c0.036,0.034,0.07,0.07,0.105,0.104   c0.037,0.035,0.072,0.072,0.109,0.108c0.034,0.034,0.068,0.069,0.102,0.102c0.035,0.037,0.071,0.073,0.106,0.108   c0.037,0.037,0.072,0.071,0.108,0.108c0.035,0.035,0.07,0.068,0.106,0.106c0.033,0.033,0.069,0.07,0.106,0.105   c0.036,0.033,0.069,0.071,0.106,0.105c0.033,0.036,0.068,0.071,0.107,0.105c0.033,0.037,0.07,0.071,0.104,0.108   c0.033,0.034,0.072,0.067,0.104,0.104c0.039,0.036,0.076,0.074,0.111,0.11c0.037,0.034,0.07,0.07,0.107,0.104   c0.035,0.037,0.068,0.071,0.105,0.104c0.031,0.036,0.07,0.069,0.102,0.104c0.035,0.036,0.072,0.07,0.105,0.108   c0.041,0.037,0.08,0.073,0.117,0.114c0.031,0.029,0.061,0.062,0.094,0.091c0.037,0.039,0.072,0.077,0.111,0.112   c0.035,0.037,0.072,0.072,0.107,0.106c0.033,0.039,0.072,0.074,0.107,0.112c0.033,0.032,0.068,0.063,0.1,0.1   c0.016,0.013,0.027,0.023,0.041,0.037C49.086,60.623,60.623,49.086,61.867,34.598z" opacity="0.2"/><path d="M32,17.213c8.155,0,14.791,6.632,14.791,14.787c0,8.154-6.636,14.789-14.791,14.789   c-8.155,0-14.789-6.635-14.789-14.789C17.21,23.845,23.844,17.213,32,17.213 M32,12.712c-10.651,0-19.29,8.636-19.29,19.288   c0,10.65,8.638,19.288,19.29,19.288c10.651,0,19.29-8.638,19.29-19.288C51.289,21.349,42.65,12.712,32,12.712L32,12.712z" fill="#FFFFFF" id="CERCHIO_1_"/></g></svg>`,
-  'X': `<?xml version="1.0" ?><!DOCTYPE svg  PUBLIC '-//W3C//DTD SVG 1.1//EN'  'http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd'><svg enable-background="new 0 0 15 15" height="15px" id="Livello_1" version="1.1" viewBox="0 0 64 64" width="64px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g><path d="M62,32c0,16.564-13.436,30-30,30C15.433,62,2,48.564,2,32C2,15.433,15.433,2,32,2C48.564,2,62,15.433,62,32   z" fill="#D2527F"/><path d="M62,32c0-0.996-0.054-1.978-0.147-2.948c-0.013-0.012-0.026-0.025-0.041-0.039   c-0.035-0.04-0.074-0.075-0.11-0.112c-0.032-0.036-0.07-0.071-0.105-0.105c-0.035-0.035-0.069-0.071-0.104-0.106   c-0.037-0.036-0.071-0.071-0.106-0.104c-0.035-0.037-0.069-0.073-0.108-0.107c-0.032-0.036-0.067-0.069-0.102-0.103   c-0.035-0.037-0.072-0.074-0.107-0.11c-0.034-0.037-0.073-0.072-0.109-0.108c-0.034-0.035-0.066-0.068-0.101-0.1   c-0.037-0.039-0.072-0.076-0.113-0.111c-0.03-0.033-0.063-0.066-0.097-0.099c-0.036-0.04-0.075-0.077-0.114-0.114   c-0.031-0.035-0.067-0.069-0.102-0.104c-0.037-0.037-0.07-0.07-0.107-0.106c-0.036-0.037-0.069-0.071-0.107-0.107   c-0.033-0.033-0.066-0.069-0.104-0.104c-0.033-0.037-0.071-0.071-0.106-0.104c-0.033-0.039-0.072-0.076-0.109-0.111   c-0.034-0.033-0.066-0.066-0.098-0.099c-0.038-0.039-0.073-0.074-0.112-0.11c-0.033-0.037-0.067-0.072-0.107-0.108   c-0.032-0.035-0.068-0.068-0.102-0.102c-0.034-0.037-0.071-0.074-0.106-0.109c-0.037-0.034-0.069-0.069-0.104-0.105   c-0.038-0.037-0.071-0.073-0.108-0.104c-0.035-0.037-0.069-0.073-0.108-0.108c-0.032-0.036-0.067-0.067-0.102-0.102   c-0.035-0.04-0.074-0.077-0.112-0.112c-0.03-0.034-0.064-0.066-0.097-0.101c-0.039-0.037-0.073-0.074-0.109-0.108   c-0.038-0.037-0.069-0.072-0.108-0.108c-0.034-0.035-0.067-0.069-0.105-0.105c-0.031-0.035-0.069-0.071-0.103-0.105   c-0.035-0.037-0.076-0.074-0.112-0.111c-0.035-0.036-0.068-0.067-0.101-0.1c-0.037-0.037-0.069-0.073-0.108-0.106   c-0.033-0.037-0.067-0.072-0.106-0.108c-0.033-0.037-0.071-0.071-0.106-0.104c-0.031-0.037-0.069-0.073-0.104-0.108   c-0.037-0.036-0.069-0.069-0.104-0.104c-0.036-0.036-0.069-0.071-0.107-0.105c-0.035-0.037-0.07-0.074-0.109-0.109   c-0.033-0.037-0.069-0.069-0.104-0.103c-0.032-0.036-0.067-0.071-0.102-0.105c-0.039-0.036-0.072-0.072-0.107-0.108   c-0.036-0.036-0.069-0.071-0.107-0.104c-0.035-0.038-0.068-0.073-0.109-0.11c-0.032-0.036-0.067-0.069-0.103-0.103   c-0.034-0.037-0.071-0.073-0.108-0.108c-0.032-0.034-0.067-0.067-0.102-0.102c-0.037-0.037-0.071-0.074-0.106-0.108   c-0.037-0.037-0.07-0.071-0.108-0.106c-0.033-0.036-0.067-0.071-0.106-0.104c-0.031-0.037-0.067-0.071-0.102-0.105   c-0.035-0.037-0.074-0.075-0.11-0.111c-0.038-0.037-0.069-0.069-0.104-0.103c-0.036-0.037-0.068-0.071-0.107-0.105   c-0.033-0.035-0.067-0.071-0.106-0.106c-0.033-0.037-0.069-0.073-0.104-0.106c-0.033-0.036-0.071-0.071-0.104-0.107   c-0.038-0.037-0.071-0.069-0.106-0.104c-0.036-0.037-0.069-0.073-0.107-0.107c-0.035-0.037-0.07-0.074-0.11-0.111   c-0.032-0.034-0.066-0.067-0.101-0.098c-0.033-0.037-0.07-0.074-0.105-0.11c-0.038-0.035-0.071-0.069-0.106-0.105   c-0.036-0.037-0.069-0.072-0.108-0.106c-0.034-0.036-0.066-0.071-0.105-0.107c-0.033-0.037-0.07-0.07-0.105-0.104   c-0.034-0.037-0.07-0.073-0.105-0.108c-0.035-0.033-0.07-0.069-0.104-0.104c-0.037-0.036-0.071-0.071-0.106-0.104   c-0.035-0.037-0.069-0.073-0.109-0.108c-0.032-0.036-0.066-0.069-0.104-0.104c-0.033-0.037-0.071-0.073-0.107-0.107   c-0.032-0.036-0.07-0.07-0.104-0.105c-0.037-0.037-0.071-0.07-0.106-0.105c-0.036-0.037-0.069-0.071-0.106-0.106   c-0.034-0.035-0.067-0.071-0.106-0.106c-0.032-0.036-0.068-0.069-0.102-0.104c-0.035-0.037-0.076-0.075-0.112-0.112   c-0.038-0.035-0.071-0.069-0.105-0.105c-0.035-0.034-0.067-0.068-0.105-0.102c-0.033-0.035-0.066-0.072-0.105-0.108   c-0.034-0.037-0.073-0.072-0.108-0.106c-0.032-0.036-0.067-0.069-0.102-0.104c-0.037-0.037-0.071-0.072-0.106-0.108   c-0.037-0.035-0.07-0.071-0.108-0.105c-0.033-0.036-0.067-0.07-0.106-0.105c-0.031-0.037-0.069-0.071-0.104-0.106   c-0.032-0.036-0.069-0.071-0.104-0.105c-0.037-0.037-0.072-0.074-0.107-0.109c-0.036-0.036-0.07-0.071-0.105-0.103   c-0.035-0.037-0.069-0.071-0.108-0.107c-0.032-0.037-0.067-0.071-0.104-0.106c-0.033-0.036-0.069-0.071-0.104-0.107   c-0.034-0.035-0.071-0.071-0.106-0.104c-0.037-0.037-0.069-0.072-0.104-0.106c-0.038-0.036-0.069-0.071-0.108-0.106   c-0.034-0.035-0.067-0.071-0.106-0.105c-0.032-0.037-0.069-0.072-0.104-0.105c-0.034-0.036-0.071-0.072-0.106-0.107   c-0.038-0.037-0.069-0.07-0.104-0.106c-0.038-0.036-0.071-0.071-0.109-0.107c-0.033-0.035-0.066-0.069-0.105-0.104   c-0.032-0.037-0.07-0.073-0.105-0.106c-0.033-0.037-0.07-0.071-0.105-0.106c-0.036-0.036-0.068-0.069-0.104-0.104   c-0.037-0.037-0.071-0.073-0.11-0.107c-0.033-0.036-0.067-0.072-0.106-0.105c-0.033-0.037-0.069-0.071-0.104-0.105   c-0.033-0.037-0.069-0.073-0.104-0.108c-0.037-0.035-0.072-0.071-0.107-0.106c-0.036-0.036-0.068-0.069-0.105-0.103   c-0.034-0.038-0.067-0.073-0.106-0.109c-0.034-0.037-0.069-0.071-0.106-0.106c-0.032-0.036-0.069-0.07-0.104-0.104   c-0.035-0.035-0.07-0.072-0.105-0.108c-0.038-0.037-0.071-0.072-0.106-0.105c-0.038-0.037-0.069-0.072-0.108-0.108   c-0.034-0.035-0.066-0.069-0.105-0.105c-0.031-0.035-0.067-0.069-0.102-0.104c-0.035-0.037-0.074-0.075-0.112-0.112   c-0.036-0.036-0.067-0.067-0.102-0.1c-0.037-0.039-0.071-0.074-0.11-0.11c-0.033-0.033-0.065-0.069-0.104-0.104   c-0.525-0.552-1.259-0.902-2.076-0.902H17.625c-1.579,0-2.875,1.295-2.875,2.874v28.751c0,0.817,0.351,1.551,0.902,2.076   c0.035,0.039,0.071,0.071,0.106,0.104c0.034,0.039,0.069,0.073,0.106,0.11c0.035,0.035,0.07,0.069,0.107,0.106   c0.034,0.033,0.066,0.069,0.103,0.102c0.033,0.035,0.069,0.072,0.107,0.106c0.036,0.039,0.069,0.072,0.104,0.106   c0.036,0.039,0.071,0.07,0.108,0.108c0.033,0.035,0.067,0.068,0.104,0.106c0.036,0.035,0.071,0.07,0.106,0.104   c0.034,0.035,0.069,0.072,0.106,0.104c0.034,0.037,0.068,0.072,0.105,0.106c0.035,0.037,0.069,0.071,0.106,0.106   c0.033,0.037,0.069,0.07,0.106,0.108c0.035,0.035,0.07,0.067,0.104,0.104c0.035,0.035,0.072,0.072,0.109,0.106   c0.033,0.033,0.066,0.069,0.102,0.102c0.036,0.039,0.071,0.072,0.107,0.106c0.034,0.04,0.07,0.074,0.108,0.111   c0.034,0.035,0.068,0.067,0.104,0.104c0.035,0.035,0.071,0.073,0.108,0.104c0.032,0.035,0.067,0.073,0.104,0.106   c0.035,0.038,0.071,0.071,0.104,0.104c0.036,0.038,0.071,0.071,0.106,0.109c0.036,0.035,0.069,0.066,0.106,0.104   c0.036,0.035,0.071,0.072,0.107,0.106c0.033,0.035,0.068,0.072,0.105,0.104c0.034,0.039,0.069,0.072,0.104,0.106   c0.036,0.037,0.07,0.07,0.107,0.106c0.034,0.035,0.068,0.069,0.105,0.106c0.034,0.035,0.07,0.072,0.105,0.106   c0.035,0.035,0.07,0.071,0.106,0.104c0.035,0.037,0.069,0.071,0.104,0.104c0.035,0.039,0.072,0.073,0.109,0.108   c0.032,0.035,0.064,0.067,0.101,0.104c0.037,0.037,0.074,0.072,0.11,0.111c0.034,0.033,0.07,0.069,0.105,0.103   c0.035,0.035,0.069,0.073,0.106,0.106c0.035,0.037,0.069,0.071,0.105,0.104c0.034,0.039,0.068,0.072,0.105,0.109   c0.034,0.034,0.069,0.067,0.104,0.104c0.037,0.037,0.074,0.076,0.111,0.109c0.033,0.035,0.069,0.071,0.104,0.104   c0.034,0.038,0.068,0.069,0.104,0.102c0.035,0.041,0.07,0.074,0.107,0.112c0.036,0.035,0.069,0.068,0.106,0.104   c0.034,0.035,0.069,0.072,0.106,0.106c0.034,0.035,0.069,0.07,0.105,0.104c0.035,0.039,0.07,0.072,0.105,0.106   c0.036,0.037,0.07,0.07,0.107,0.106c0.034,0.035,0.067,0.069,0.104,0.106c0.035,0.033,0.071,0.071,0.105,0.104   c0.034,0.035,0.07,0.071,0.107,0.104c0.035,0.039,0.069,0.074,0.106,0.108c0.033,0.036,0.066,0.067,0.102,0.104   c0.036,0.038,0.074,0.073,0.112,0.112c0.034,0.034,0.068,0.066,0.102,0.102c0.036,0.035,0.071,0.071,0.108,0.105   c0.034,0.035,0.067,0.072,0.104,0.105c0.036,0.039,0.071,0.071,0.106,0.105c0.034,0.039,0.069,0.072,0.106,0.108   c0.036,0.035,0.069,0.068,0.104,0.106c0.037,0.036,0.074,0.074,0.111,0.107c0.031,0.032,0.063,0.066,0.096,0.096   c0.037,0.042,0.076,0.078,0.112,0.113c0.036,0.039,0.069,0.071,0.106,0.107c0.036,0.035,0.068,0.068,0.105,0.106   c0.035,0.033,0.07,0.071,0.106,0.104c0.033,0.033,0.069,0.071,0.104,0.104c0.035,0.039,0.071,0.073,0.106,0.106   c0.035,0.039,0.071,0.072,0.108,0.108c0.033,0.035,0.065,0.067,0.102,0.104c0.037,0.036,0.074,0.075,0.109,0.109   c0.033,0.033,0.069,0.071,0.106,0.103c0.034,0.039,0.068,0.073,0.104,0.106c0.035,0.038,0.069,0.071,0.106,0.108   c0.033,0.034,0.067,0.067,0.104,0.104c0.037,0.037,0.072,0.072,0.109,0.109c0.034,0.033,0.069,0.071,0.104,0.104   c0.036,0.035,0.071,0.072,0.108,0.105c0.034,0.038,0.067,0.07,0.102,0.105c0.035,0.035,0.071,0.07,0.108,0.108   c0.035,0.035,0.07,0.068,0.105,0.106c0.036,0.035,0.071,0.07,0.107,0.104c0.034,0.035,0.068,0.071,0.105,0.104   c0.034,0.039,0.07,0.072,0.105,0.104c0.035,0.039,0.069,0.073,0.106,0.11c0.035,0.035,0.069,0.067,0.104,0.104   c0.035,0.035,0.071,0.073,0.108,0.104c0.033,0.035,0.069,0.074,0.107,0.108c0.034,0.037,0.069,0.071,0.104,0.103   c0.034,0.038,0.066,0.07,0.102,0.104c0.037,0.038,0.071,0.073,0.111,0.112c0.034,0.035,0.071,0.072,0.106,0.104   c0.033,0.035,0.069,0.071,0.104,0.104c0.036,0.038,0.07,0.071,0.104,0.105c0.034,0.037,0.07,0.07,0.105,0.106   c0.037,0.038,0.074,0.073,0.114,0.114c0.03,0.03,0.063,0.063,0.096,0.092c0.035,0.039,0.072,0.078,0.11,0.113   c0.036,0.035,0.069,0.072,0.106,0.104c0.034,0.039,0.069,0.072,0.106,0.106c0.032,0.037,0.068,0.07,0.105,0.108   c0.035,0.035,0.07,0.067,0.104,0.104c0.035,0.035,0.071,0.072,0.108,0.106c0.034,0.033,0.067,0.071,0.104,0.103   c0.035,0.039,0.069,0.073,0.105,0.106c0.034,0.038,0.07,0.071,0.105,0.108c0.036,0.035,0.07,0.067,0.106,0.105   c0.035,0.035,0.071,0.072,0.108,0.105c0.033,0.035,0.067,0.073,0.104,0.106c0.035,0.038,0.071,0.071,0.104,0.104   c0.036,0.038,0.071,0.071,0.106,0.109c0.036,0.035,0.069,0.066,0.106,0.104c0.036,0.036,0.071,0.074,0.108,0.107   c0.033,0.034,0.067,0.069,0.104,0.102c0.033,0.039,0.069,0.073,0.105,0.106c0.034,0.039,0.07,0.072,0.107,0.108   c0.033,0.035,0.067,0.069,0.103,0.104c0.037,0.037,0.073,0.074,0.109,0.109c0.036,0.034,0.069,0.071,0.106,0.104   c0.034,0.037,0.068,0.071,0.105,0.104c0.033,0.039,0.069,0.073,0.106,0.108c0.033,0.035,0.069,0.069,0.104,0.106   c0.035,0.035,0.071,0.069,0.106,0.104c0.034,0.035,0.069,0.073,0.106,0.105c0.035,0.036,0.07,0.074,0.107,0.109   c0.015,0.014,0.026,0.027,0.041,0.042C30.021,61.946,31.004,62,32,62C48.564,62,62,48.564,62,32z" opacity="0.2"/><path d="M44.937,19.063v25.873H19.063V19.063H44.937 M46.376,14.75H17.625c-1.579,0-2.875,1.295-2.875,2.874v28.751   c0,1.58,1.296,2.876,2.875,2.876h28.751c1.58,0,2.876-1.296,2.876-2.876V17.625C49.252,16.046,47.956,14.75,46.376,14.75   L46.376,14.75z" fill="#FFFFFF"/></g></svg>`,
-  'Y': `<?xml version="1.0" ?><!DOCTYPE svg  PUBLIC '-//W3C//DTD SVG 1.1//EN'  'http://www.w3.org/Graphics/SVG/1.1/DTD/svg11.dtd'><svg enable-background="new 0 0 15 15" height="15px" id="Livello_1" version="1.1" viewBox="0 0 64 64" width="64px" xml:space="preserve" xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink"><g><path d="M62,32c0,16.564-13.434,30-30,30C15.432,62,2,48.564,2,32C2,15.432,15.432,2,32,2C48.566,2,62,15.432,62,32   z" fill="#65C6BB"/><path d="M60.584,41.105c-0.004-0.005-0.006-0.008-0.009-0.011c-0.034-0.041-0.069-0.069-0.106-0.104   c-0.033-0.044-0.073-0.071-0.107-0.11c-0.031-0.038-0.07-0.065-0.104-0.104c-0.036-0.04-0.072-0.072-0.107-0.107   c-0.035-0.04-0.07-0.069-0.106-0.104c-0.034-0.04-0.071-0.071-0.108-0.105c-0.029-0.043-0.067-0.069-0.102-0.107   c-0.033-0.041-0.074-0.068-0.107-0.108c-0.036-0.04-0.072-0.069-0.105-0.104c-0.037-0.04-0.072-0.07-0.106-0.105   c-0.034-0.038-0.069-0.07-0.108-0.104c-0.031-0.044-0.071-0.072-0.105-0.107c-0.033-0.044-0.071-0.071-0.105-0.108   c-0.036-0.04-0.074-0.067-0.107-0.105c-0.035-0.039-0.07-0.07-0.105-0.104c-0.035-0.04-0.07-0.071-0.106-0.106   c-0.034-0.041-0.069-0.069-0.106-0.105c-0.033-0.043-0.071-0.071-0.105-0.109c-0.033-0.039-0.072-0.066-0.105-0.104   c-0.038-0.04-0.072-0.072-0.107-0.107c-0.035-0.038-0.068-0.068-0.106-0.102c-0.034-0.042-0.067-0.073-0.106-0.108   c-0.031-0.043-0.071-0.07-0.105-0.108c-0.033-0.04-0.072-0.067-0.105-0.106c-0.036-0.038-0.07-0.069-0.105-0.105   c-0.037-0.039-0.072-0.07-0.108-0.105c-0.032-0.039-0.069-0.07-0.106-0.104c-0.033-0.044-0.071-0.071-0.105-0.106   c-0.033-0.042-0.07-0.07-0.105-0.109c-0.036-0.037-0.07-0.066-0.104-0.104c-0.037-0.04-0.072-0.072-0.109-0.107   c-0.035-0.039-0.07-0.07-0.106-0.104c-0.034-0.044-0.069-0.072-0.106-0.107c-0.033-0.041-0.069-0.068-0.102-0.105   c-0.037-0.04-0.074-0.07-0.109-0.109c-0.036-0.039-0.072-0.07-0.105-0.105c-0.035-0.038-0.07-0.069-0.108-0.104   c-0.034-0.04-0.067-0.071-0.104-0.106c-0.033-0.042-0.071-0.07-0.107-0.108c-0.031-0.04-0.072-0.067-0.105-0.104   c-0.034-0.04-0.07-0.071-0.105-0.107c-0.037-0.037-0.07-0.068-0.106-0.104c-0.034-0.038-0.069-0.069-0.106-0.104   c-0.033-0.044-0.073-0.073-0.107-0.109c-0.033-0.042-0.07-0.068-0.105-0.107c-0.036-0.038-0.07-0.067-0.104-0.104   c-0.037-0.038-0.072-0.071-0.108-0.106c-0.034-0.038-0.069-0.071-0.107-0.104c-0.032-0.044-0.069-0.071-0.106-0.104   c-0.031-0.044-0.07-0.071-0.104-0.108c-0.035-0.04-0.072-0.07-0.107-0.108c-0.034-0.04-0.072-0.071-0.105-0.106   c-0.035-0.038-0.07-0.069-0.108-0.104c-0.034-0.04-0.067-0.069-0.104-0.106c-0.033-0.042-0.071-0.069-0.105-0.108   c-0.033-0.038-0.074-0.067-0.107-0.104c-0.034-0.04-0.07-0.071-0.104-0.104c-0.037-0.04-0.072-0.073-0.11-0.108   c-0.032-0.038-0.067-0.067-0.104-0.103c-0.033-0.044-0.071-0.073-0.107-0.108c-0.031-0.042-0.07-0.069-0.104-0.106   c-0.036-0.04-0.072-0.069-0.107-0.108c-0.035-0.038-0.07-0.067-0.106-0.103c-0.034-0.038-0.069-0.071-0.106-0.106   c-0.033-0.044-0.07-0.071-0.107-0.104c-0.031-0.044-0.07-0.071-0.104-0.11c-0.035-0.038-0.072-0.067-0.105-0.106   c-0.037-0.04-0.074-0.071-0.107-0.106c-0.035-0.038-0.07-0.069-0.106-0.103c-0.032-0.042-0.069-0.071-0.106-0.106   c-0.033-0.042-0.071-0.069-0.105-0.11c-0.033-0.036-0.07-0.065-0.104-0.104c-0.038-0.039-0.074-0.071-0.109-0.106   c-0.035-0.038-0.07-0.069-0.106-0.103c-0.034-0.04-0.069-0.073-0.108-0.108c-0.031-0.042-0.067-0.068-0.102-0.106   c-0.035-0.04-0.074-0.071-0.109-0.106c-0.034-0.04-0.068-0.067-0.102-0.104c-0.037-0.04-0.074-0.071-0.11-0.108   c-0.034-0.041-0.069-0.071-0.106-0.106c-0.033-0.04-0.07-0.069-0.105-0.105c-0.031-0.041-0.072-0.07-0.105-0.105   c-0.034-0.041-0.072-0.07-0.105-0.107c-0.035-0.039-0.07-0.069-0.105-0.106c-0.037-0.041-0.072-0.07-0.107-0.104   c-0.033-0.041-0.068-0.07-0.107-0.107c-0.033-0.039-0.071-0.069-0.105-0.106c-0.033-0.04-0.07-0.069-0.104-0.105   c-0.038-0.04-0.074-0.071-0.109-0.108c-0.035-0.039-0.07-0.067-0.104-0.103c-0.034-0.041-0.071-0.072-0.11-0.11   c-0.031-0.037-0.067-0.066-0.102-0.103c-0.035-0.042-0.074-0.072-0.109-0.108c-0.032-0.039-0.068-0.067-0.102-0.104   c-0.037-0.04-0.074-0.07-0.11-0.107c-0.032-0.041-0.067-0.071-0.104-0.107c-0.035-0.039-0.072-0.069-0.107-0.105   c-0.031-0.04-0.07-0.069-0.104-0.104c-0.036-0.041-0.074-0.072-0.107-0.11c-0.037-0.038-0.07-0.066-0.105-0.103   c-0.037-0.041-0.07-0.071-0.107-0.105c-0.033-0.041-0.068-0.07-0.105-0.107c-0.033-0.039-0.073-0.07-0.107-0.107   c-0.033-0.041-0.07-0.07-0.104-0.106c-0.038-0.039-0.072-0.067-0.105-0.104c-0.037-0.041-0.072-0.071-0.11-0.106   c-0.032-0.04-0.069-0.069-0.104-0.106c-0.031-0.039-0.071-0.068-0.105-0.105c-0.033-0.042-0.074-0.073-0.109-0.109   c-0.032-0.04-0.07-0.069-0.102-0.105c-0.037-0.038-0.072-0.067-0.106-0.104c-0.034-0.042-0.071-0.072-0.108-0.109   c-0.035-0.04-0.072-0.07-0.107-0.107c-0.031-0.04-0.072-0.069-0.105-0.104c-0.034-0.04-0.068-0.068-0.102-0.104   c-0.037-0.04-0.072-0.072-0.109-0.109c-0.037-0.042-0.07-0.071-0.107-0.106c-0.033-0.041-0.068-0.07-0.105-0.107   c-0.033-0.039-0.072-0.068-0.105-0.104c-0.033-0.041-0.072-0.07-0.105-0.107c-0.037-0.039-0.072-0.069-0.105-0.106   c-0.037-0.041-0.07-0.07-0.108-0.105c-0.034-0.041-0.069-0.07-0.108-0.107c-0.029-0.038-0.069-0.067-0.104-0.104   c-0.031-0.041-0.072-0.071-0.105-0.106c-0.036-0.041-0.07-0.071-0.105-0.108c-0.037-0.039-0.072-0.068-0.11-0.105   c-0.03-0.04-0.065-0.069-0.104-0.106c-0.031-0.039-0.071-0.069-0.105-0.105c-0.033-0.042-0.07-0.072-0.105-0.107   c-0.036-0.041-0.072-0.069-0.104-0.105c-0.037-0.039-0.072-0.069-0.109-0.106c-0.035-0.041-0.07-0.071-0.107-0.107   c-0.033-0.04-0.068-0.07-0.105-0.106c-0.033-0.04-0.07-0.069-0.104-0.105c-0.035-0.041-0.072-0.07-0.107-0.107   c-0.037-0.04-0.072-0.07-0.107-0.107c-0.035-0.041-0.068-0.069-0.107-0.104c-0.033-0.041-0.066-0.071-0.104-0.108   c-0.033-0.039-0.073-0.069-0.107-0.106c-0.033-0.04-0.072-0.069-0.105-0.105c-0.036-0.04-0.07-0.07-0.105-0.106   c-0.037-0.04-0.07-0.069-0.108-0.106c-0.032-0.041-0.067-0.07-0.104-0.107c-0.033-0.039-0.071-0.068-0.107-0.105   c-0.031-0.041-0.07-0.07-0.105-0.105c-0.036-0.042-0.072-0.072-0.105-0.109c-0.035-0.039-0.072-0.068-0.107-0.104   c-0.033-0.041-0.068-0.069-0.104-0.105c-0.035-0.041-0.072-0.072-0.109-0.109c-0.033-0.039-0.07-0.068-0.104-0.104   c-0.035-0.041-0.072-0.071-0.107-0.107c-0.035-0.039-0.07-0.068-0.104-0.104c-0.037-0.042-0.072-0.072-0.111-0.107   c-0.033-0.041-0.066-0.07-0.104-0.106c-0.033-0.04-0.072-0.07-0.107-0.107c-0.033-0.041-0.072-0.07-0.105-0.105   c-0.035-0.04-0.07-0.069-0.104-0.105c-0.035-0.041-0.072-0.07-0.109-0.107c-0.033-0.041-0.068-0.072-0.107-0.109   c-0.031-0.038-0.069-0.067-0.104-0.102c-0.033-0.041-0.072-0.071-0.105-0.107c-0.036-0.042-0.072-0.071-0.107-0.109   c-0.035-0.038-0.07-0.066-0.106-0.102c-0.034-0.042-0.069-0.072-0.108-0.108c-0.031-0.041-0.068-0.07-0.105-0.107   c-0.031-0.039-0.07-0.069-0.104-0.104c-0.035-0.041-0.072-0.07-0.105-0.107c-0.037-0.039-0.072-0.069-0.107-0.106   c-0.035-0.041-0.07-0.07-0.109-0.105c-0.031-0.041-0.066-0.071-0.104-0.107c-0.033-0.04-0.072-0.069-0.105-0.104   c-0.033-0.041-0.072-0.071-0.105-0.107c-0.037-0.04-0.072-0.07-0.107-0.106c-0.035-0.04-0.07-0.069-0.107-0.104   c-0.033-0.041-0.068-0.071-0.105-0.108c-0.033-0.039-0.071-0.069-0.105-0.106c-0.033-0.04-0.072-0.07-0.107-0.105   c-0.034-0.041-0.07-0.07-0.104-0.106c-0.037-0.041-0.072-0.071-0.109-0.107c-0.033-0.04-0.068-0.069-0.105-0.105   c-0.033-0.04-0.07-0.07-0.107-0.107c-0.031-0.039-0.07-0.068-0.104-0.104c-0.034-0.041-0.07-0.07-0.104-0.106   c-0.035-0.04-0.072-0.071-0.107-0.108c-0.037-0.041-0.072-0.07-0.109-0.105c-0.031-0.041-0.068-0.071-0.105-0.108   c-0.033-0.04-0.072-0.069-0.105-0.105c-0.033-0.04-0.07-0.07-0.104-0.106c-0.039-0.04-0.074-0.07-0.109-0.107   c-0.035-0.039-0.07-0.068-0.105-0.103c-0.033-0.041-0.07-0.071-0.107-0.108c-0.031-0.039-0.07-0.068-0.104-0.105   c-0.035-0.042-0.074-0.072-0.109-0.108c-0.035-0.04-0.068-0.068-0.102-0.104c-0.037-0.041-0.074-0.07-0.11-0.108   c-0.034-0.041-0.069-0.071-0.106-0.107c-0.033-0.04-0.069-0.068-0.104-0.104c-0.033-0.041-0.072-0.071-0.107-0.107   c-0.035-0.04-0.072-0.069-0.105-0.106c-0.035-0.039-0.07-0.069-0.105-0.106c-0.037-0.041-0.07-0.07-0.107-0.105   c-0.033-0.041-0.07-0.071-0.107-0.108c-0.033-0.04-0.072-0.069-0.105-0.104c-0.033-0.041-0.07-0.07-0.104-0.107   c-0.039-0.039-0.072-0.069-0.105-0.105c-0.037-0.042-0.074-0.071-0.111-0.107c-0.033-0.04-0.068-0.069-0.105-0.106   c-0.031-0.04-0.07-0.069-0.104-0.105c-0.035-0.042-0.072-0.071-0.107-0.107c-0.035-0.04-0.072-0.07-0.105-0.106   c-0.035-0.039-0.07-0.068-0.107-0.104c-0.031-0.041-0.068-0.071-0.105-0.108c-0.035-0.04-0.072-0.07-0.107-0.107   c-0.031-0.04-0.068-0.068-0.104-0.103c-0.036-0.041-0.072-0.072-0.107-0.11c-0.037-0.039-0.07-0.068-0.105-0.104   c-0.037-0.041-0.07-0.07-0.107-0.105c-0.033-0.041-0.068-0.07-0.105-0.106c-0.033-0.041-0.074-0.07-0.107-0.107   c-0.031-0.041-0.07-0.07-0.104-0.107c-0.039-0.039-0.072-0.069-0.105-0.106c-0.037-0.041-0.072-0.07-0.109-0.104   c-0.033-0.041-0.07-0.071-0.107-0.108c-0.031-0.04-0.07-0.069-0.104-0.105c-0.033-0.04-0.072-0.069-0.104-0.104   c-0.037-0.041-0.072-0.071-0.107-0.109c-0.037-0.04-0.074-0.069-0.11-0.105c-0.03-0.04-0.065-0.069-0.103-0.106   c-0.033-0.04-0.074-0.07-0.109-0.108c-0.031-0.04-0.068-0.069-0.104-0.104c-0.037-0.041-0.072-0.07-0.105-0.107   c-0.035-0.04-0.072-0.07-0.107-0.106c-0.035-0.04-0.07-0.07-0.107-0.105c-0.033-0.041-0.068-0.071-0.105-0.107   c-0.033-0.04-0.072-0.07-0.107-0.106c-0.031-0.041-0.07-0.071-0.105-0.108c-0.037-0.039-0.07-0.067-0.104-0.104   c-0.035-0.041-0.07-0.07-0.109-0.105c-0.031-0.042-0.066-0.071-0.105-0.108c-0.031-0.041-0.072-0.07-0.105-0.106   c-0.031-0.04-0.07-0.069-0.102-0.104c-0.039-0.041-0.074-0.072-0.109-0.109c-0.037-0.04-0.072-0.069-0.109-0.105   c-0.031-0.04-0.068-0.07-0.105-0.106c-0.033-0.039-0.072-0.068-0.105-0.105c-0.031-0.041-0.07-0.07-0.105-0.105   c-0.037-0.041-0.072-0.071-0.107-0.108c-0.033-0.039-0.068-0.067-0.105-0.104c-0.033-0.041-0.07-0.071-0.107-0.107   c-0.033-0.04-0.068-0.069-0.105-0.105c-0.033-0.04-0.07-0.069-0.104-0.106c-0.035-0.04-0.072-0.07-0.105-0.107   c-0.037-0.04-0.074-0.069-0.107-0.106c-0.035-0.041-0.07-0.07-0.109-0.105c-0.033-0.041-0.066-0.07-0.104-0.107   c-0.033-0.039-0.072-0.068-0.105-0.105c-0.033-0.041-0.074-0.071-0.107-0.106c-0.356-0.418-0.769-0.63-1.184-0.63   c-0.542,0-1.085,0.359-1.497,1.073L12.557,45.329c-0.526,0.912-0.435,1.714,0.125,2.177c0.036,0.039,0.067,0.074,0.107,0.109   c0.035,0.035,0.066,0.07,0.106,0.106c0.035,0.034,0.067,0.071,0.106,0.104c0.034,0.034,0.066,0.074,0.105,0.105   c0.035,0.036,0.068,0.075,0.107,0.107c0.035,0.037,0.067,0.074,0.107,0.105c0.033,0.037,0.064,0.072,0.104,0.106   c0.036,0.035,0.068,0.072,0.106,0.106c0.036,0.035,0.068,0.075,0.108,0.107c0.035,0.034,0.065,0.072,0.104,0.104   c0.036,0.039,0.067,0.075,0.107,0.106c0.034,0.037,0.065,0.073,0.105,0.108c0.035,0.035,0.068,0.071,0.107,0.107   c0.035,0.033,0.066,0.072,0.106,0.104c0.033,0.035,0.065,0.074,0.105,0.105c0.036,0.038,0.067,0.074,0.106,0.105   c0.035,0.039,0.066,0.074,0.106,0.108c0.035,0.035,0.067,0.07,0.105,0.106c0.036,0.035,0.068,0.072,0.108,0.105   c0.034,0.034,0.065,0.074,0.105,0.105c0.035,0.035,0.067,0.074,0.107,0.105c0.035,0.039,0.066,0.073,0.106,0.106   c0.034,0.038,0.066,0.073,0.106,0.108c0.035,0.034,0.067,0.071,0.105,0.105c0.036,0.035,0.068,0.075,0.109,0.107   c0.035,0.034,0.065,0.071,0.104,0.102c0.036,0.041,0.068,0.076,0.107,0.108c0.034,0.037,0.065,0.073,0.105,0.108   c0.035,0.033,0.067,0.07,0.106,0.105c0.036,0.034,0.067,0.073,0.106,0.105c0.034,0.035,0.066,0.074,0.106,0.105   c0.035,0.038,0.066,0.073,0.104,0.106c0.037,0.036,0.07,0.073,0.111,0.108c0.035,0.035,0.064,0.069,0.104,0.104   c0.035,0.035,0.066,0.072,0.106,0.104c0.033,0.037,0.065,0.076,0.106,0.108c0.035,0.035,0.067,0.074,0.106,0.104   c0.035,0.039,0.066,0.074,0.106,0.106c0.034,0.037,0.065,0.072,0.106,0.108c0.035,0.035,0.067,0.071,0.104,0.106   c0.036,0.033,0.069,0.074,0.109,0.104c0.035,0.036,0.066,0.074,0.105,0.104c0.036,0.041,0.068,0.076,0.107,0.108   c0.033,0.037,0.064,0.072,0.104,0.106c0.036,0.035,0.068,0.073,0.108,0.107c0.035,0.034,0.065,0.073,0.104,0.104   c0.035,0.036,0.066,0.074,0.107,0.106c0.035,0.039,0.067,0.075,0.106,0.107c0.035,0.036,0.066,0.073,0.107,0.107   c0.035,0.033,0.066,0.07,0.105,0.104c0.035,0.035,0.066,0.072,0.105,0.104c0.035,0.037,0.068,0.077,0.109,0.108   c0.034,0.035,0.066,0.074,0.104,0.104c0.035,0.037,0.066,0.074,0.105,0.106c0.035,0.035,0.066,0.072,0.106,0.106   c0.036,0.037,0.068,0.073,0.106,0.108c0.035,0.033,0.066,0.074,0.106,0.104c0.035,0.034,0.067,0.074,0.107,0.106   c0.035,0.039,0.066,0.072,0.105,0.104c0.034,0.038,0.066,0.075,0.106,0.109c0.036,0.035,0.067,0.073,0.107,0.106   c0.034,0.033,0.066,0.074,0.105,0.104c0.035,0.034,0.066,0.074,0.106,0.104c0.036,0.039,0.067,0.076,0.105,0.107   c0.036,0.038,0.068,0.075,0.109,0.109c0.035,0.033,0.065,0.07,0.104,0.104c0.035,0.033,0.066,0.072,0.106,0.104   c0.034,0.036,0.066,0.076,0.107,0.108c0.035,0.035,0.066,0.074,0.105,0.104c0.035,0.037,0.066,0.074,0.106,0.106   c0.035,0.035,0.065,0.072,0.105,0.106c0.035,0.035,0.067,0.071,0.106,0.106c0.035,0.035,0.067,0.076,0.107,0.106   c0.035,0.034,0.067,0.074,0.106,0.104c0.035,0.039,0.067,0.074,0.106,0.105c0.033,0.038,0.064,0.073,0.104,0.107   c0.037,0.035,0.069,0.073,0.109,0.108c0.035,0.033,0.066,0.074,0.106,0.104c0.034,0.034,0.065,0.074,0.105,0.104   c0.036,0.039,0.066,0.074,0.105,0.105c0.035,0.038,0.067,0.073,0.107,0.109c0.035,0.035,0.066,0.071,0.105,0.106   c0.035,0.033,0.068,0.072,0.107,0.106c0.033,0.034,0.065,0.072,0.104,0.103c0.036,0.039,0.068,0.076,0.106,0.107   c0.035,0.038,0.067,0.073,0.107,0.107c0.035,0.035,0.066,0.072,0.106,0.108c0.036,0.033,0.067,0.069,0.105,0.103   c0.035,0.035,0.067,0.074,0.106,0.105c0.036,0.037,0.067,0.075,0.106,0.107c0.035,0.037,0.067,0.074,0.107,0.105   c0.034,0.038,0.065,0.073,0.106,0.109c0.035,0.033,0.067,0.069,0.104,0.104c0.036,0.036,0.067,0.075,0.107,0.107   c0.035,0.034,0.066,0.074,0.107,0.104c0.035,0.039,0.066,0.074,0.105,0.105c0.035,0.038,0.067,0.073,0.107,0.109   c0.035,0.035,0.066,0.07,0.104,0.104c0.036,0.035,0.069,0.074,0.109,0.106c0.033,0.036,0.065,0.074,0.105,0.104   c0.035,0.039,0.067,0.074,0.105,0.107c0.035,0.038,0.067,0.073,0.107,0.107c0.035,0.035,0.064,0.07,0.104,0.106   c0.036,0.034,0.067,0.071,0.106,0.104c0.035,0.035,0.068,0.076,0.108,0.108c0.035,0.034,0.065,0.071,0.103,0.103   c0.036,0.038,0.068,0.074,0.108,0.107c0.034,0.037,0.066,0.073,0.106,0.107c0.036,0.035,0.068,0.071,0.106,0.106   c0.035,0.034,0.066,0.073,0.105,0.104c0.035,0.037,0.066,0.075,0.106,0.107c0.036,0.038,0.067,0.074,0.106,0.105   c0.034,0.037,0.066,0.073,0.106,0.108c0.036,0.035,0.067,0.071,0.105,0.105c0.035,0.034,0.066,0.074,0.106,0.105   c0.034,0.036,0.066,0.075,0.107,0.105c0.035,0.039,0.066,0.074,0.103,0.106c0.036,0.036,0.068,0.073,0.109,0.108   c0.035,0.035,0.065,0.071,0.105,0.106c0.036,0.034,0.066,0.07,0.106,0.104c0.034,0.035,0.066,0.074,0.106,0.105   c0.035,0.035,0.067,0.074,0.105,0.105c0.036,0.039,0.068,0.074,0.109,0.107c0.033,0.037,0.065,0.073,0.105,0.107   c0.035,0.034,0.065,0.07,0.104,0.105c0.036,0.033,0.068,0.074,0.109,0.105c0.034,0.036,0.065,0.074,0.105,0.106   c0.036,0.038,0.067,0.074,0.106,0.105c0.034,0.037,0.066,0.073,0.106,0.108c0.035,0.033,0.067,0.071,0.105,0.104   c0.036,0.035,0.067,0.075,0.108,0.107c0.027,0.027,0.055,0.056,0.084,0.085C28.407,61.811,30.18,62,32,62   C45.39,62,56.727,53.219,60.584,41.105z" opacity="0.2"/><path d="M32,20.566l13.157,22.861H18.842L32,20.566 M32,13.073c-0.542,0-1.085,0.359-1.497,1.073L12.557,45.329   c-0.824,1.427-0.146,2.598,1.503,2.598h35.882c1.65,0,2.324-1.171,1.502-2.598L33.494,14.146   C33.084,13.432,32.542,13.073,32,13.073L32,13.073z" fill="#FFFFFF"/></g></svg>`
-}
